@@ -1,23 +1,23 @@
 [package]
 name = "czkawka_cli"
-version = "5.1.0"
+version = "6.0.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
 edition = "2021"
-rust-version = "1.65.0"
+rust-version = "1.67.1"
 description = "CLI frontend of Czkawka"
 license = "MIT"
 homepage = "https://github.com/qarmin/czkawka"
 repository = "https://github.com/qarmin/czkawka"
 
 [dependencies]
-clap = { version = "4.1", features = ["derive"] }
+clap = { version = "4.3", features = ["derive"] }
 
 # For enum types
-image_hasher = "1.1.2"
+image_hasher = "1.2"
 
 [dependencies.czkawka_core]
 path = "../czkawka_core"
-version = "5.1.0"
+version = "6.0.0"
 features = []
 
 [features]
