@@ -1,9 +1,9 @@
 [package]
 name = "czkawka_core"
-version = "5.1.0"
+version = "6.0.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
 edition = "2021"
-rust-version = "1.65.0"
+rust-version = "1.67.1"
 description = "Core of Czkawka app"
 license = "MIT"
 homepage = "https://github.com/qarmin/czkawka"
@@ -11,69 +11,74 @@ repository = "https://github.com/qarmin/czkawka"
 
 
 [dependencies]
-humansize = "2.1.3"
-rayon = "1.6.1"
-crossbeam-channel = "0.5.6"
+humansize = "2.1"
+rayon = "1.7"
+crossbeam-channel = "0.5"
 
 # For saving/loading config files to specific directories
-directories-next = "2.0.0"
+directories-next = "2.0"
 
 # Needed by similar images
-image_hasher = "1.1.2"
-bk-tree = "0.4.0"
-image = "0.24.5"
-hamming = "0.1.3"
+image_hasher = "1.2"
+bk-tree = "0.5"
+image = "0.24"
+hamming = "0.1"
 
 # Needed by same music
-bitflags = "1.3.2"
-lofty = "0.11.0"
+bitflags = "2.3"
+lofty = "0.14"
 
 # Futures - needed by async progress sender
-futures = "0.3.26"
+futures = "0.3.28"
 
 # Needed by broken files
-zip = { version = "0.6.4", features = ["aes-crypto", "bzip2", "deflate", "time"], default-features = false }
-audio_checker = "0.1.0"
-pdf = "0.8.0"
+zip = { version = "0.6", features = ["aes-crypto", "bzip2", "deflate", "time"], default-features = false }
+audio_checker = "0.1"
+pdf = "0.8"
+
+# Needed by audio similarity feature
+rusty-chromaprint = "0.1"
+symphonia = { version = "0.5", features = ["all"] }
 
 # Hashes for duplicate files
-blake3 = "1.3.3"
-crc32fast = "1.3.2"
-xxhash-rust = { version = "0.8.6", features = ["xxh3"] }
+blake3 = "1.4"
+crc32fast = "1.3"
+xxhash-rust = { version = "0.8", features = ["xxh3"] }
 
-tempfile = "3.3.0"
+tempfile = "3.6"
 
 # Video Duplicates
-vid_dup_finder_lib = "0.1.1"
-ffmpeg_cmdline_utils = "0.1.2"
+vid_dup_finder_lib = "0.1"
+ffmpeg_cmdline_utils = "0.1"
 
 # Saving/Loading Cache
 serde = "1.0"
-bincode = "1.3.3"
+bincode = "1.3"
 serde_json = "1.0"
 
 # Language
-i18n-embed = { version = "0.13.8", features = ["fluent-system", "desktop-requester"] }
-i18n-embed-fl = "0.6.5"
-rust-embed = "6.4.2"
-once_cell = "1.17.1"
+i18n-embed = { version = "0.13", features = ["fluent-system", "desktop-requester"] }
+i18n-embed-fl = "0.6"
+rust-embed = "6.7"
+once_cell = "1.18"
 
 # Raw image files
-rawloader = "0.37.1"
-imagepipe = "0.5.0"
+rawloader = "0.37"
+imagepipe = "0.5"
 
 # Checking for invalid extensions
-mime_guess = "2.0.4"
-infer = "0.12.0"
+mime_guess = "2.0"
+infer = "0.13"
 
-num_cpus = "1.15.0"
+num_cpus = "1.15"
 
 # Heif/Heic
-libheif-rs = { version = "0.16.0", optional = true }
-anyhow = { version = "1.0", optional = true }
+libheif-rs = { version = "=0.18.0", optional = true } # Do not upgrade now, since Ubuntu 22.04 not works with newer version
+libheif-sys = { version = "=1.14.2", optional = true } # 1.14.3 brake compilation on Ubuntu 22.04
+anyhow = { version = "1.0" }
 
-state="0.5.3"
+state = "0.6"
 
 [features]
 default = []
-heif = ["dep:libheif-rs", "dep:anyhow"]
+heif = ["dep:libheif-rs", "dep:libheif-sys"]
