@@ -1,65 +1,65 @@
 [package]
 name = "czkawka_gui"
-version = "5.1.0"
+version = "6.0.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
 edition = "2021"
-rust-version = "1.65.0"
+rust-version = "1.67.1"
 description = "GTK frontend of Czkawka"
 license = "MIT"
 homepage = "https://github.com/qarmin/czkawka"
 repository = "https://github.com/qarmin/czkawka"
 
 [dependencies]
-gdk4 = "0.6.0"
-glib = "0.17.1"
+gdk4 = "0.6"
+glib = "0.17"
 
-humansize = "2.1.3"
-chrono = "0.4.23"
+humansize = "2.1"
+chrono = "0.4.26"
 
 # Used for sending stop signal across threads
-crossbeam-channel = "0.5.6"
+crossbeam-channel = "0.5.8"
 
 # To get information about progress
-futures = "0.3.26"
+futures = "0.3.28"
 
 # For saving/loading config files to specific directories
-directories-next = "2.0.0"
+directories-next = "2.0"
 
 # For opening files
-open = "3.2.0"
+open = "4.1"
 
 # To get image preview
-image = "0.24.5"
+image = "0.24"
 
 # To be able to use custom select
-regex = "1.7.1"
+regex = "1.8"
 
 # To get image_hasher types
-image_hasher = "1.1.2"
+image_hasher = "1.2"
 
 # Move files to trash
-trash = "3.0.1"
+trash = "3.0"
 
-# For moving files(why std::fs doesn't have such features)
-fs_extra = "1.3.0"
+# For moving files(why std::fs doesn't have such features?)
+fs_extra = "1.3"
 
 # Language
-i18n-embed = { version = "0.13.8", features = ["fluent-system", "desktop-requester"] }
-i18n-embed-fl = "0.6.5"
-rust-embed = "6.4.2"
-once_cell = "1.17.1"
+i18n-embed = { version = "0.13", features = ["fluent-system", "desktop-requester"] }
+i18n-embed-fl = "0.6"
+rust-embed = "6.7"
+once_cell = "1.18"
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3.9", features = ["combaseapi", "objbase", "shobjidl_core", "windef", "winerror", "wtypesbase", "winuser"] }
 
 [dependencies.gtk4]
-version = "0.6.1"
+version = "0.6"
 default-features = false
 features = ["v4_6"]
 
 [dependencies.czkawka_core]
 path = "../czkawka_core"
-version = "5.1.0"
+version = "6.0.0"
 features = []
 
 [features]
