@@ -9,19 +9,19 @@ use std::io::{BufReader, BufWriter};
 #[cfg(target_family = "unix")]
 use std::os::unix::fs::MetadataExt;
 use std::path::{Path, PathBuf};
-use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-use std::sync::Arc;
-use std::thread::sleep;
-use std::time::{Duration, SystemTime};
-use std::{fs, mem, thread};
+use std::sync::atomic::Ordering;
+
+use std::{fs, mem};
 
 use crossbeam_channel::Receiver;
+use futures::channel::mpsc::UnboundedSender;
 use humansize::format_size;
 use humansize::BINARY;
 use rayon::prelude::*;
+use xxhash_rust::xxh3::Xxh3;
 
-use crate::common::{open_cache_folder, Common, LOOP_DURATION};
-use crate::common_dir_traversal::{CheckingMethod, DirTraversalBuilder, DirTraversalResult, FileEntry, ProgressData};
+use crate::common::{open_cache_folder, prepare_thread_handler_common, send_info_and_wait_for_ending_all_threads};
+use crate::common_dir_traversal::{CheckingMethod, DirTraversalBuilder, DirTraversalResult, FileEntry, ProgressData, ToolType};
 use crate::common_directory::Directories;
 use crate::common_extensions::Extensions;
 use crate::common_items::ExcludedItems;
@@ -44,7 +44,7 @@ impl HashType {
         match self {
             HashType::Blake3 => Box::new(blake3::Hasher::new()),
             HashType::Crc32 => Box::new(crc32fast::Hasher::new()),
-            HashType::Xxh3 => Box::new(xxhash_rust::xxh3::Xxh3::new()),
+            HashType::Xxh3 => Box::new(Xxh3::new()),
         }
     }
 }
@@ -67,6 +67,8 @@ pub struct Info {
     pub number_of_duplicated_files_by_hash: usize,
     pub number_of_groups_by_name: usize,
     pub number_of_duplicated_files_by_name: usize,
+    pub number_of_groups_by_size_name: usize,
+    pub number_of_duplicated_files_by_size_name: usize,
     pub lost_space_by_size: u64,
     pub lost_space_by_hash: u64,
 }
@@ -79,13 +81,16 @@ impl Info {
 }
 
 pub struct DuplicateFinder {
+    tool_type: ToolType,
     text_messages: Messages,
     information: Info,
-    files_with_identical_names: BTreeMap<String, Vec<FileEntry>>,                            // File Size, File Entry
-    files_with_identical_size: BTreeMap<u64, Vec<FileEntry>>,                                // File Size, File Entry
-    files_with_identical_hashes: BTreeMap<u64, Vec<Vec<FileEntry>>>,                         // File Size, next grouped by file size, next grouped by hash
-    files_with_identical_names_referenced: BTreeMap<String, (FileEntry, Vec<FileEntry>)>,    // File Size, File Entry
-    files_with_identical_size_referenced: BTreeMap<u64, (FileEntry, Vec<FileEntry>)>,        // File Size, File Entry
+    files_with_identical_names: BTreeMap<String, Vec<FileEntry>>,                         // File Size, File Entry
+    files_with_identical_size_names: BTreeMap<(u64, String), Vec<FileEntry>>,             // File (Size, Name), File Entry
+    files_with_identical_size: BTreeMap<u64, Vec<FileEntry>>,                             // File Size, File Entry
+    files_with_identical_hashes: BTreeMap<u64, Vec<Vec<FileEntry>>>,                      // File Size, next grouped by file size, next grouped by hash
+    files_with_identical_names_referenced: BTreeMap<String, (FileEntry, Vec<FileEntry>)>, // File Size, File Entry
+    files_with_identical_size_names_referenced: BTreeMap<(u64, String), (FileEntry, Vec<FileEntry>)>, // File (Size, Name), File Entry
+    files_with_identical_size_referenced: BTreeMap<u64, (FileEntry, Vec<FileEntry>)>,     // File Size, File Entry
     files_with_identical_hashes_referenced: BTreeMap<u64, Vec<(FileEntry, Vec<FileEntry>)>>, // File Size, next grouped by file size, next grouped by hash
     directories: Directories,
     allowed_extensions: Extensions,
@@ -112,12 +117,15 @@ impl DuplicateFinder {
     #[must_use]
     pub fn new() -> Self {
         Self {
+            tool_type: ToolType::Duplicate,
             text_messages: Messages::new(),
             information: Info::new(),
             files_with_identical_names: Default::default(),
             files_with_identical_size: Default::default(),
+            files_with_identical_size_names: Default::default(),
             files_with_identical_hashes: Default::default(),
             files_with_identical_names_referenced: Default::default(),
+            files_with_identical_size_names_referenced: Default::default(),
             files_with_identical_size_referenced: Default::default(),
             files_with_identical_hashes_referenced: Default::default(),
             recursive_search: true,
@@ -142,36 +150,40 @@ impl DuplicateFinder {
         }
     }
 
-    pub fn find_duplicates(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) {
+    pub fn find_duplicates(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) {
         self.directories.optimize_directories(self.recursive_search, &mut self.text_messages);
         self.use_reference_folders = !self.directories.reference_directories.is_empty();
 
         match self.check_method {
             CheckingMethod::Name => {
-                if !self.check_files_name(stop_receiver, progress_sender) {
-                    self.stopped_search = true;
+                self.stopped_search = !self.check_files_name(stop_receiver, progress_sender); // TODO restore this to name
+                if self.stopped_search {
+                    return;
+                }
+            }
+            CheckingMethod::SizeName => {
+                self.stopped_search = !self.check_files_size_name(stop_receiver, progress_sender);
+                if self.stopped_search {
                     return;
                 }
             }
             CheckingMethod::Size => {
-                if !self.check_files_size(stop_receiver, progress_sender) {
-                    self.stopped_search = true;
+                self.stopped_search = !self.check_files_size(stop_receiver, progress_sender);
+                if self.stopped_search {
                     return;
                 }
             }
             CheckingMethod::Hash => {
-                if !self.check_files_size(stop_receiver, progress_sender) {
-                    self.stopped_search = true;
+                self.stopped_search = !self.check_files_size(stop_receiver, progress_sender);
+                if self.stopped_search {
                     return;
                 }
-                if !self.check_files_hash(stop_receiver, progress_sender) {
-                    self.stopped_search = true;
+                self.stopped_search = !self.check_files_hash(stop_receiver, progress_sender);
+                if self.stopped_search {
                     return;
                 }
             }
-            CheckingMethod::None => {
-                panic!();
-            }
+            _ => panic!(),
         }
         self.delete_files();
         self.debug_print();
@@ -219,6 +231,11 @@ impl DuplicateFinder {
     #[must_use]
     pub const fn get_files_sorted_by_size(&self) -> &BTreeMap<u64, Vec<FileEntry>> {
         &self.files_with_identical_size
+    }
+
+    #[must_use]
+    pub const fn get_files_sorted_by_size_name(&self) -> &BTreeMap<(u64, String), Vec<FileEntry>> {
+        &self.files_with_identical_size_names
     }
 
     #[must_use]
@@ -319,7 +336,12 @@ impl DuplicateFinder {
         &self.files_with_identical_size_referenced
     }
 
-    fn check_files_name(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) -> bool {
+    #[must_use]
+    pub fn get_files_with_identical_size_names_referenced(&self) -> &BTreeMap<(u64, String), (FileEntry, Vec<FileEntry>)> {
+        &self.files_with_identical_size_names_referenced
+    }
+
+    fn check_files_name(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
         let group_by_func = if self.case_sensitive_name_comparison {
             |fe: &FileEntry| fe.path.file_name().unwrap().to_string_lossy().to_string()
         } else {
@@ -341,11 +363,7 @@ impl DuplicateFinder {
             .build()
             .run();
         match result {
-            DirTraversalResult::SuccessFiles {
-                start_time,
-                grouped_file_entries,
-                warnings,
-            } => {
+            DirTraversalResult::SuccessFiles { grouped_file_entries, warnings } => {
                 self.files_with_identical_names = grouped_file_entries;
                 self.text_messages.warnings.extend(warnings);
 
@@ -361,21 +379,11 @@ impl DuplicateFinder {
 
                 // Reference - only use in size, because later hash will be counted differently
                 if self.use_reference_folders {
-                    let mut btree_map = Default::default();
-                    mem::swap(&mut self.files_with_identical_names, &mut btree_map);
-                    let reference_directories = self.directories.reference_directories.clone();
-                    let vec = btree_map
+                    let vec = mem::take(&mut self.files_with_identical_names)
                         .into_iter()
-                        .filter_map(|(_size, vec_file_entry)| {
-                            let mut files_from_referenced_folders = Vec::new();
-                            let mut normal_files = Vec::new();
-                            for file_entry in vec_file_entry {
-                                if reference_directories.iter().any(|e| file_entry.path.starts_with(e)) {
-                                    files_from_referenced_folders.push(file_entry);
-                                } else {
-                                    normal_files.push(file_entry);
-                                }
-                            }
+                        .filter_map(|(_name, vec_file_entry)| {
+                            let (mut files_from_referenced_folders, normal_files): (Vec<_>, Vec<_>) =
+                                vec_file_entry.into_iter().partition(|e| self.directories.is_in_referenced_directory(e.get_path()));
 
                             if files_from_referenced_folders.is_empty() || normal_files.is_empty() {
                                 None
@@ -388,20 +396,8 @@ impl DuplicateFinder {
                         self.files_with_identical_names_referenced.insert(fe.path.to_string_lossy().to_string(), (fe, vec_fe));
                     }
                 }
+                self.calculate_name_stats();
 
-                if self.use_reference_folders {
-                    for (_fe, vector) in self.files_with_identical_names_referenced.values() {
-                        self.information.number_of_duplicated_files_by_name += vector.len();
-                        self.information.number_of_groups_by_name += 1;
-                    }
-                } else {
-                    for vector in self.files_with_identical_names.values() {
-                        self.information.number_of_duplicated_files_by_name += vector.len() - 1;
-                        self.information.number_of_groups_by_name += 1;
-                    }
-                }
-
-                Common::print_time(start_time, SystemTime::now(), "check_files_name");
                 true
             }
             DirTraversalResult::SuccessFolders { .. } => {
@@ -411,9 +407,106 @@ impl DuplicateFinder {
         }
     }
 
+    fn calculate_name_stats(&mut self) {
+        if self.use_reference_folders {
+            for (_fe, vector) in self.files_with_identical_names_referenced.values() {
+                self.information.number_of_duplicated_files_by_name += vector.len();
+                self.information.number_of_groups_by_name += 1;
+            }
+        } else {
+            for vector in self.files_with_identical_names.values() {
+                self.information.number_of_duplicated_files_by_name += vector.len() - 1;
+                self.information.number_of_groups_by_name += 1;
+            }
+        }
+    }
+
+    fn check_files_size_name(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
+        let group_by_func = if self.case_sensitive_name_comparison {
+            |fe: &FileEntry| (fe.size, fe.path.file_name().unwrap().to_string_lossy().to_string())
+        } else {
+            |fe: &FileEntry| (fe.size, fe.path.file_name().unwrap().to_string_lossy().to_lowercase())
+        };
+
+        let result = DirTraversalBuilder::new()
+            .root_dirs(self.directories.included_directories.clone())
+            .group_by(group_by_func)
+            .stop_receiver(stop_receiver)
+            .progress_sender(progress_sender)
+            .checking_method(CheckingMethod::Name)
+            .directories(self.directories.clone())
+            .allowed_extensions(self.allowed_extensions.clone())
+            .excluded_items(self.excluded_items.clone())
+            .recursive_search(self.recursive_search)
+            .minimal_file_size(self.minimal_file_size)
+            .maximal_file_size(self.maximal_file_size)
+            .build()
+            .run();
+        match result {
+            DirTraversalResult::SuccessFiles { grouped_file_entries, warnings } => {
+                self.files_with_identical_size_names = grouped_file_entries;
+                self.text_messages.warnings.extend(warnings);
+
+                // Create new BTreeMap without single size entries(files have not duplicates)
+                let mut new_map: BTreeMap<(u64, String), Vec<FileEntry>> = Default::default();
+
+                for (name_size, vector) in &self.files_with_identical_size_names {
+                    if vector.len() > 1 {
+                        new_map.insert(name_size.clone(), vector.clone());
+                    }
+                }
+                self.files_with_identical_size_names = new_map;
+
+                // Reference - only use in size, because later hash will be counted differently
+                if self.use_reference_folders {
+                    let vec = mem::take(&mut self.files_with_identical_size_names)
+                        .into_iter()
+                        .filter_map(|(_size, vec_file_entry)| {
+                            let (mut files_from_referenced_folders, normal_files): (Vec<_>, Vec<_>) =
+                                vec_file_entry.into_iter().partition(|e| self.directories.is_in_referenced_directory(e.get_path()));
+
+                            if files_from_referenced_folders.is_empty() || normal_files.is_empty() {
+                                None
+                            } else {
+                                Some((files_from_referenced_folders.pop().unwrap(), normal_files))
+                            }
+                        })
+                        .collect::<Vec<(FileEntry, Vec<FileEntry>)>>();
+                    for (fe, vec_fe) in vec {
+                        self.files_with_identical_size_names_referenced
+                            .insert((fe.size, fe.path.to_string_lossy().to_string()), (fe, vec_fe));
+                    }
+                }
+                self.calculate_size_name_stats();
+
+                true
+            }
+            DirTraversalResult::SuccessFolders { .. } => {
+                unreachable!()
+            }
+            DirTraversalResult::Stopped => false,
+        }
+    }
+
+    fn calculate_size_name_stats(&mut self) {
+        if self.use_reference_folders {
+            for ((size, _name), (_fe, vector)) in &self.files_with_identical_size_names_referenced {
+                self.information.number_of_duplicated_files_by_size_name += vector.len();
+                self.information.number_of_groups_by_size_name += 1;
+                self.information.lost_space_by_size += (vector.len() as u64) * size;
+            }
+        } else {
+            for ((size, _name), vector) in &self.files_with_identical_size_names {
+                self.information.number_of_duplicated_files_by_size_name += vector.len() - 1;
+                self.information.number_of_groups_by_size_name += 1;
+                self.information.lost_space_by_size += (vector.len() as u64 - 1) * size;
+            }
+        }
+    }
+
     /// Read file length and puts it to different boxes(each for different lengths)
     /// If in box is only 1 result, then it is removed
-    fn check_files_size(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) -> bool {
+    fn check_files_size(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
         let max_stage = match self.check_method {
             CheckingMethod::Size => 0,
             CheckingMethod::Hash => 2,
@@ -435,17 +528,12 @@ impl DuplicateFinder {
             .build()
             .run();
         match result {
-            DirTraversalResult::SuccessFiles {
-                start_time,
-                grouped_file_entries,
-                warnings,
-            } => {
+            DirTraversalResult::SuccessFiles { grouped_file_entries, warnings } => {
                 self.files_with_identical_size = grouped_file_entries;
                 self.text_messages.warnings.extend(warnings);
 
                 // Create new BTreeMap without single size entries(files have not duplicates)
-                let mut old_map: BTreeMap<u64, Vec<FileEntry>> = Default::default();
-                mem::swap(&mut old_map, &mut self.files_with_identical_size);
+                let old_map: BTreeMap<u64, Vec<FileEntry>> = mem::take(&mut self.files_with_identical_size);
 
                 for (size, vec) in old_map {
                     if vec.len() <= 1 {
@@ -459,51 +547,9 @@ impl DuplicateFinder {
                     }
                 }
 
-                // Reference - only use in size, because later hash will be counted differently
-                if self.use_reference_folders && self.check_method == CheckingMethod::Size {
-                    let mut btree_map = Default::default();
-                    mem::swap(&mut self.files_with_identical_size, &mut btree_map);
-                    let reference_directories = self.directories.reference_directories.clone();
-                    let vec = btree_map
-                        .into_iter()
-                        .filter_map(|(_size, vec_file_entry)| {
-                            let mut files_from_referenced_folders = Vec::new();
-                            let mut normal_files = Vec::new();
-                            for file_entry in vec_file_entry {
-                                if reference_directories.iter().any(|e| file_entry.path.starts_with(e)) {
-                                    files_from_referenced_folders.push(file_entry);
-                                } else {
-                                    normal_files.push(file_entry);
-                                }
-                            }
+                self.filter_reference_folders_by_size();
+                self.calculate_size_stats();
 
-                            if files_from_referenced_folders.is_empty() || normal_files.is_empty() {
-                                None
-                            } else {
-                                Some((files_from_referenced_folders.pop().unwrap(), normal_files))
-                            }
-                        })
-                        .collect::<Vec<(FileEntry, Vec<FileEntry>)>>();
-                    for (fe, vec_fe) in vec {
-                        self.files_with_identical_size_referenced.insert(fe.size, (fe, vec_fe));
-                    }
-                }
-
-                if self.use_reference_folders {
-                    for (size, (_fe, vector)) in &self.files_with_identical_size_referenced {
-                        self.information.number_of_duplicated_files_by_size += vector.len();
-                        self.information.number_of_groups_by_size += 1;
-                        self.information.lost_space_by_size += (vector.len() as u64) * size;
-                    }
-                } else {
-                    for (size, vector) in &self.files_with_identical_size {
-                        self.information.number_of_duplicated_files_by_size += vector.len() - 1;
-                        self.information.number_of_groups_by_size += 1;
-                        self.information.lost_space_by_size += (vector.len() as u64 - 1) * size;
-                    }
-                }
-
-                Common::print_time(start_time, SystemTime::now(), "check_files_size");
                 true
             }
             DirTraversalResult::SuccessFolders { .. } => {
@@ -513,260 +559,300 @@ impl DuplicateFinder {
         }
     }
 
-    /// The slowest checking type, which must be applied after checking for size
-    fn check_files_hash(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) -> bool {
-        assert_eq!(self.check_method, CheckingMethod::Hash);
-
-        let check_type = Arc::new(self.hash_type);
-
-        let start_time: SystemTime = SystemTime::now();
-        let check_was_stopped = AtomicBool::new(false); // Used for breaking from GUI and ending check thread
-        let mut pre_checked_map: BTreeMap<u64, Vec<FileEntry>> = Default::default();
-
-        //// PROGRESS THREAD START
-        let progress_thread_run = Arc::new(AtomicBool::new(true));
-
-        let atomic_file_counter = Arc::new(AtomicUsize::new(0));
-
-        let progress_thread_handle = if let Some(progress_sender) = progress_sender {
-            let progress_send = progress_sender.clone();
-            let progress_thread_run = progress_thread_run.clone();
-            let atomic_file_counter = atomic_file_counter.clone();
-            let files_to_check = self.files_with_identical_size.values().map(Vec::len).sum();
-            let checking_method = self.check_method;
-            thread::spawn(move || loop {
-                progress_send
-                    .unbounded_send(ProgressData {
-                        checking_method,
-                        current_stage: 1,
-                        max_stage: 2,
-                        entries_checked: atomic_file_counter.load(Ordering::Relaxed),
-                        entries_to_check: files_to_check,
-                    })
-                    .unwrap();
-                if !progress_thread_run.load(Ordering::Relaxed) {
-                    break;
-                }
-                sleep(Duration::from_millis(LOOP_DURATION as u64));
-            })
+    fn calculate_size_stats(&mut self) {
+        if self.use_reference_folders {
+            for (size, (_fe, vector)) in &self.files_with_identical_size_referenced {
+                self.information.number_of_duplicated_files_by_size += vector.len();
+                self.information.number_of_groups_by_size += 1;
+                self.information.lost_space_by_size += (vector.len() as u64) * size;
+            }
         } else {
-            thread::spawn(|| {})
-        };
-
-        //// PROGRESS THREAD END
-
-        ///////////////////////////////////////////////////////////////////////////// PREHASHING START
-        {
-            let loaded_hash_map;
-            let mut records_already_cached: BTreeMap<u64, Vec<FileEntry>> = Default::default();
-            let mut non_cached_files_to_check: BTreeMap<u64, Vec<FileEntry>> = Default::default();
-
-            // Cache algorithm
-            // - Load data from cache
-            // - Convert from BT<u64,Vec<FileEntry>> to BT<String,FileEntry>
-            // - Save to proper values
-            if self.use_prehash_cache {
-                loaded_hash_map = match load_hashes_from_file(&mut self.text_messages, self.delete_outdated_cache, &self.hash_type, true) {
-                    Some(t) => t,
-                    None => Default::default(),
-                };
-
-                let mut loaded_hash_map2: BTreeMap<String, FileEntry> = Default::default();
-                for vec_file_entry in loaded_hash_map.values() {
-                    for file_entry in vec_file_entry {
-                        loaded_hash_map2.insert(file_entry.path.to_string_lossy().to_string(), file_entry.clone());
-                    }
-                }
-
-                #[allow(clippy::if_same_then_else)]
-                for vec_file_entry in self.files_with_identical_size.values() {
-                    for file_entry in vec_file_entry {
-                        let name = file_entry.path.to_string_lossy().to_string();
-                        if !loaded_hash_map2.contains_key(&name) {
-                            // If loaded data doesn't contains current image info
-                            non_cached_files_to_check.entry(file_entry.size).or_insert_with(Vec::new).push(file_entry.clone());
-                        } else if file_entry.size != loaded_hash_map2.get(&name).unwrap().size || file_entry.modified_date != loaded_hash_map2.get(&name).unwrap().modified_date {
-                            // When size or modification date of image changed, then it is clear that is different image
-                            non_cached_files_to_check.entry(file_entry.size).or_insert_with(Vec::new).push(file_entry.clone());
-                        } else {
-                            // Checking may be omitted when already there is entry with same size and modification date
-                            records_already_cached.entry(file_entry.size).or_insert_with(Vec::new).push(file_entry.clone());
-                        }
-                    }
-                }
-            } else {
-                loaded_hash_map = Default::default();
-                mem::swap(&mut self.files_with_identical_size, &mut non_cached_files_to_check);
+            for (size, vector) in &self.files_with_identical_size {
+                self.information.number_of_duplicated_files_by_size += vector.len() - 1;
+                self.information.number_of_groups_by_size += 1;
+                self.information.lost_space_by_size += (vector.len() as u64 - 1) * size;
             }
+        }
+    }
 
-            #[allow(clippy::type_complexity)]
-            let pre_hash_results: Vec<(u64, BTreeMap<String, Vec<FileEntry>>, Vec<String>)> = non_cached_files_to_check
-                .par_iter()
-                .map(|(size, vec_file_entry)| {
-                    let mut hashmap_with_hash: BTreeMap<String, Vec<FileEntry>> = Default::default();
-                    let mut errors: Vec<String> = Vec::new();
-                    let mut buffer = [0u8; 1024 * 2];
+    /// This step check for references, only when checking for size.
+    /// This is needed, because later reference folders looks for hashes, not size
+    fn filter_reference_folders_by_size(&mut self) {
+        if self.use_reference_folders && self.check_method == CheckingMethod::Size {
+            let vec = mem::take(&mut self.files_with_identical_size)
+                .into_iter()
+                .filter_map(|(_size, vec_file_entry)| {
+                    let (mut files_from_referenced_folders, normal_files): (Vec<_>, Vec<_>) =
+                        vec_file_entry.into_iter().partition(|e| self.directories.is_in_referenced_directory(e.get_path()));
 
-                    atomic_file_counter.fetch_add(vec_file_entry.len(), Ordering::Relaxed);
-                    for file_entry in vec_file_entry {
-                        if stop_receiver.is_some() && stop_receiver.unwrap().try_recv().is_ok() {
-                            check_was_stopped.store(true, Ordering::Relaxed);
-                            return None;
-                        }
-                        match hash_calculation(&mut buffer, file_entry, &check_type, 0) {
-                            Ok(hash_string) => {
-                                hashmap_with_hash.entry(hash_string.clone()).or_insert_with(Vec::new).push(file_entry.clone());
-                            }
-                            Err(s) => errors.push(s),
-                        }
+                    if files_from_referenced_folders.is_empty() || normal_files.is_empty() {
+                        None
+                    } else {
+                        Some((files_from_referenced_folders.pop().unwrap(), normal_files))
                     }
-                    Some((*size, hashmap_with_hash, errors))
                 })
-                .while_some()
-                .collect();
+                .collect::<Vec<(FileEntry, Vec<FileEntry>)>>();
+            for (fe, vec_fe) in vec {
+                self.files_with_identical_size_referenced.insert(fe.size, (fe, vec_fe));
+            }
+        }
+    }
 
-            // End thread which send info to gui
-            progress_thread_run.store(false, Ordering::Relaxed);
-            progress_thread_handle.join().unwrap();
+    fn prehash_load_cache_at_start(&mut self) -> (BTreeMap<u64, Vec<FileEntry>>, BTreeMap<u64, Vec<FileEntry>>, BTreeMap<u64, Vec<FileEntry>>) {
+        // Cache algorithm
+        // - Load data from cache
+        // - Convert from BT<u64,Vec<FileEntry>> to BT<String,FileEntry>
+        // - Save to proper values
+        let loaded_hash_map;
+        let mut records_already_cached: BTreeMap<u64, Vec<FileEntry>> = Default::default();
+        let mut non_cached_files_to_check: BTreeMap<u64, Vec<FileEntry>> = Default::default();
 
-            // Check if user aborted search(only from GUI)
-            if check_was_stopped.load(Ordering::Relaxed) {
-                return false;
+        if self.use_prehash_cache {
+            loaded_hash_map = match load_hashes_from_file(&mut self.text_messages, self.delete_outdated_cache, &self.hash_type, true) {
+                Some(t) => t,
+                None => Default::default(),
+            };
+
+            let mut loaded_hash_map2: BTreeMap<String, FileEntry> = Default::default();
+            for vec_file_entry in loaded_hash_map.values() {
+                for file_entry in vec_file_entry {
+                    loaded_hash_map2.insert(file_entry.path.to_string_lossy().to_string(), file_entry.clone());
+                }
             }
 
-            // Add data from cache
-            for (size, vec_file_entry) in &records_already_cached {
-                pre_checked_map.entry(*size).or_insert_with(Vec::new).append(&mut vec_file_entry.clone());
+            #[allow(clippy::if_same_then_else)]
+            for vec_file_entry in self.files_with_identical_size.values() {
+                for file_entry in vec_file_entry {
+                    let name = file_entry.path.to_string_lossy().to_string();
+                    if !loaded_hash_map2.contains_key(&name) {
+                        // If loaded data doesn't contains current image info
+                        non_cached_files_to_check.entry(file_entry.size).or_insert_with(Vec::new).push(file_entry.clone());
+                    } else if file_entry.size != loaded_hash_map2.get(&name).unwrap().size || file_entry.modified_date != loaded_hash_map2.get(&name).unwrap().modified_date {
+                        // When size or modification date of image changed, then it is clear that is different image
+                        non_cached_files_to_check.entry(file_entry.size).or_insert_with(Vec::new).push(file_entry.clone());
+                    } else {
+                        // Checking may be omitted when already there is entry with same size and modification date
+                        records_already_cached.entry(file_entry.size).or_insert_with(Vec::new).push(file_entry.clone());
+                    }
+                }
             }
+        } else {
+            loaded_hash_map = Default::default();
+            mem::swap(&mut self.files_with_identical_size, &mut non_cached_files_to_check);
+        }
+        (loaded_hash_map, records_already_cached, non_cached_files_to_check)
+    }
 
-            // Check results
-            for (size, hash_map, errors) in &pre_hash_results {
-                self.text_messages.warnings.append(&mut errors.clone());
-                for vec_file_entry in hash_map.values() {
-                    if vec_file_entry.len() > 1 {
-                        pre_checked_map.entry(*size).or_insert_with(Vec::new).append(&mut vec_file_entry.clone());
+    fn prehash_save_cache_at_exit(&mut self, loaded_hash_map: BTreeMap<u64, Vec<FileEntry>>, pre_hash_results: &Vec<(u64, BTreeMap<String, Vec<FileEntry>>, Vec<String>)>) {
+        if self.use_prehash_cache {
+            // All results = records already cached + computed results
+            let mut save_cache_to_hashmap: BTreeMap<String, FileEntry> = Default::default();
+
+            for (size, vec_file_entry) in loaded_hash_map {
+                if size >= self.minimal_prehash_cache_file_size {
+                    for file_entry in vec_file_entry {
+                        save_cache_to_hashmap.insert(file_entry.path.to_string_lossy().to_string(), file_entry.clone());
                     }
                 }
             }
 
-            if self.use_prehash_cache {
-                // All results = records already cached + computed results
-                let mut save_cache_to_hashmap: BTreeMap<String, FileEntry> = Default::default();
-
-                for (size, vec_file_entry) in loaded_hash_map {
-                    if size >= self.minimal_prehash_cache_file_size {
+            for (size, hash_map, _errors) in pre_hash_results {
+                if *size >= self.minimal_prehash_cache_file_size {
+                    for vec_file_entry in hash_map.values() {
                         for file_entry in vec_file_entry {
                             save_cache_to_hashmap.insert(file_entry.path.to_string_lossy().to_string(), file_entry.clone());
                         }
                     }
                 }
+            }
 
-                for (size, hash_map, _errors) in &pre_hash_results {
-                    if *size >= self.minimal_prehash_cache_file_size {
-                        for vec_file_entry in hash_map.values() {
-                            for file_entry in vec_file_entry {
-                                save_cache_to_hashmap.insert(file_entry.path.to_string_lossy().to_string(), file_entry.clone());
-                            }
+            save_hashes_to_file(&save_cache_to_hashmap, &mut self.text_messages, &self.hash_type, true, self.minimal_prehash_cache_file_size);
+        }
+    }
+
+    fn prehashing(
+        &mut self,
+        stop_receiver: Option<&Receiver<()>>,
+        progress_sender: Option<&UnboundedSender<ProgressData>>,
+        pre_checked_map: &mut BTreeMap<u64, Vec<FileEntry>>,
+    ) -> Option<()> {
+        let check_type = self.hash_type;
+        let (progress_thread_handle, progress_thread_run, atomic_counter, check_was_stopped) = prepare_thread_handler_common(
+            progress_sender,
+            1,
+            2,
+            self.files_with_identical_size.values().map(Vec::len).sum(),
+            self.check_method,
+            self.tool_type,
+        );
+
+        let (loaded_hash_map, records_already_cached, non_cached_files_to_check) = self.prehash_load_cache_at_start();
+
+        #[allow(clippy::type_complexity)]
+        let pre_hash_results: Vec<(u64, BTreeMap<String, Vec<FileEntry>>, Vec<String>)> = non_cached_files_to_check
+            .par_iter()
+            .map(|(size, vec_file_entry)| {
+                let mut hashmap_with_hash: BTreeMap<String, Vec<FileEntry>> = Default::default();
+                let mut errors: Vec<String> = Vec::new();
+                let mut buffer = [0u8; 1024 * 2];
+
+                atomic_counter.fetch_add(vec_file_entry.len(), Ordering::Relaxed);
+                if stop_receiver.is_some() && stop_receiver.unwrap().try_recv().is_ok() {
+                    check_was_stopped.store(true, Ordering::Relaxed);
+                    return None;
+                }
+                for file_entry in vec_file_entry {
+                    match hash_calculation(&mut buffer, file_entry, &check_type, 0) {
+                        Ok(hash_string) => {
+                            hashmap_with_hash.entry(hash_string.clone()).or_insert_with(Vec::new).push(file_entry.clone());
                         }
+                        Err(s) => errors.push(s),
                     }
                 }
+                Some((*size, hashmap_with_hash, errors))
+            })
+            .while_some()
+            .collect();
 
-                save_hashes_to_file(&save_cache_to_hashmap, &mut self.text_messages, &self.hash_type, true, self.minimal_prehash_cache_file_size);
+        send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
+
+        // Check if user aborted search(only from GUI)
+        if check_was_stopped.load(Ordering::Relaxed) {
+            return None;
+        }
+
+        // Add data from cache
+        for (size, vec_file_entry) in &records_already_cached {
+            pre_checked_map.entry(*size).or_insert_with(Vec::new).append(&mut vec_file_entry.clone());
+        }
+
+        // Check results
+        for (size, hash_map, errors) in &pre_hash_results {
+            self.text_messages.warnings.append(&mut errors.clone());
+            for vec_file_entry in hash_map.values() {
+                if vec_file_entry.len() > 1 {
+                    pre_checked_map.entry(*size).or_insert_with(Vec::new).append(&mut vec_file_entry.clone());
+                }
             }
         }
 
-        ///////////////////////////////////////////////////////////////////////////// PREHASHING END
+        self.prehash_save_cache_at_exit(loaded_hash_map, &pre_hash_results);
 
-        Common::print_time(start_time, SystemTime::now(), "check_files_hash - prehash");
-        let start_time: SystemTime = SystemTime::now();
+        Some(())
+    }
 
-        /////////////////////////
+    fn full_hashing_load_cache_at_start(
+        &mut self,
+        mut pre_checked_map: BTreeMap<u64, Vec<FileEntry>>,
+    ) -> (BTreeMap<u64, Vec<FileEntry>>, BTreeMap<u64, Vec<FileEntry>>, BTreeMap<u64, Vec<FileEntry>>) {
+        let loaded_hash_map;
+        let mut records_already_cached: BTreeMap<u64, Vec<FileEntry>> = Default::default();
+        let mut non_cached_files_to_check: BTreeMap<u64, Vec<FileEntry>> = Default::default();
 
-        //// PROGRESS THREAD START
-        let progress_thread_run = Arc::new(AtomicBool::new(true));
+        if self.use_cache {
+            loaded_hash_map = match load_hashes_from_file(&mut self.text_messages, self.delete_outdated_cache, &self.hash_type, false) {
+                Some(t) => t,
+                None => Default::default(),
+            };
 
-        let atomic_file_counter = Arc::new(AtomicUsize::new(0));
+            for (size, vec_file_entry) in pre_checked_map {
+                #[allow(clippy::collapsible_if)]
+                if !loaded_hash_map.contains_key(&size) {
+                    // If loaded data doesn't contains current info
+                    non_cached_files_to_check.insert(size, vec_file_entry);
+                } else {
+                    let loaded_vec_file_entry = loaded_hash_map.get(&size).unwrap();
 
-        let progress_thread_handle = if let Some(progress_sender) = progress_sender {
-            let progress_send = progress_sender.clone();
-            let progress_thread_run = progress_thread_run.clone();
-            let atomic_file_counter = atomic_file_counter.clone();
-            let files_to_check = pre_checked_map.values().map(Vec::len).sum();
-            let checking_method = self.check_method;
-            thread::spawn(move || loop {
-                progress_send
-                    .unbounded_send(ProgressData {
-                        checking_method,
-                        current_stage: 2,
-                        max_stage: 2,
-                        entries_checked: atomic_file_counter.load(Ordering::Relaxed),
-                        entries_to_check: files_to_check,
-                    })
-                    .unwrap();
-                if !progress_thread_run.load(Ordering::Relaxed) {
-                    break;
-                }
-                sleep(Duration::from_millis(LOOP_DURATION as u64));
-            })
-        } else {
-            thread::spawn(|| {})
-        };
-
-        //// PROGRESS THREAD END
-
-        ///////////////////////////////////////////////////////////////////////////// HASHING START
-        {
-            #[allow(clippy::type_complexity)]
-            let mut full_hash_results: Vec<(u64, BTreeMap<String, Vec<FileEntry>>, Vec<String>)>;
-
-            let loaded_hash_map;
-
-            let mut records_already_cached: BTreeMap<u64, Vec<FileEntry>> = Default::default();
-            let mut non_cached_files_to_check: BTreeMap<u64, Vec<FileEntry>> = Default::default();
-
-            if self.use_cache {
-                loaded_hash_map = match load_hashes_from_file(&mut self.text_messages, self.delete_outdated_cache, &self.hash_type, false) {
-                    Some(t) => t,
-                    None => Default::default(),
-                };
-
-                for (size, vec_file_entry) in pre_checked_map {
-                    #[allow(clippy::collapsible_if)]
-                    if !loaded_hash_map.contains_key(&size) {
-                        // If loaded data doesn't contains current info
-                        non_cached_files_to_check.insert(size, vec_file_entry);
-                    } else {
-                        let loaded_vec_file_entry = loaded_hash_map.get(&size).unwrap();
-
-                        for file_entry in vec_file_entry {
-                            let mut found: bool = false;
-                            for loaded_file_entry in loaded_vec_file_entry {
-                                if file_entry.path == loaded_file_entry.path && file_entry.modified_date == loaded_file_entry.modified_date {
-                                    records_already_cached.entry(file_entry.size).or_insert_with(Vec::new).push(loaded_file_entry.clone());
-                                    found = true;
-                                    break;
-                                }
+                    for file_entry in vec_file_entry {
+                        let mut found: bool = false;
+                        for loaded_file_entry in loaded_vec_file_entry {
+                            if file_entry.path == loaded_file_entry.path && file_entry.modified_date == loaded_file_entry.modified_date {
+                                records_already_cached.entry(file_entry.size).or_insert_with(Vec::new).push(loaded_file_entry.clone());
+                                found = true;
+                                break;
                             }
+                        }
 
-                            if !found {
-                                non_cached_files_to_check.entry(file_entry.size).or_insert_with(Vec::new).push(file_entry);
-                            }
+                        if !found {
+                            non_cached_files_to_check.entry(file_entry.size).or_insert_with(Vec::new).push(file_entry);
                         }
                     }
                 }
-            } else {
-                loaded_hash_map = Default::default();
-                mem::swap(&mut pre_checked_map, &mut non_cached_files_to_check);
             }
+        } else {
+            loaded_hash_map = Default::default();
+            mem::swap(&mut pre_checked_map, &mut non_cached_files_to_check);
+        }
+        (loaded_hash_map, records_already_cached, non_cached_files_to_check)
+    }
 
-            full_hash_results = non_cached_files_to_check
+    fn full_hashing_save_cache_at_exit(
+        &mut self,
+        records_already_cached: BTreeMap<u64, Vec<FileEntry>>,
+        full_hash_results: &mut Vec<(u64, BTreeMap<String, Vec<FileEntry>>, Vec<String>)>,
+        loaded_hash_map: BTreeMap<u64, Vec<FileEntry>>,
+    ) {
+        if !self.use_cache {
+            return;
+        }
+        'main: for (size, vec_file_entry) in records_already_cached {
+            // Check if size already exists, if exists we must to change it outside because cannot have mut and non mut reference to full_hash_results
+            for (full_size, full_hashmap, _errors) in &mut (*full_hash_results) {
+                if size == *full_size {
+                    for file_entry in vec_file_entry {
+                        full_hashmap.entry(file_entry.hash.clone()).or_insert_with(Vec::new).push(file_entry);
+                    }
+                    continue 'main;
+                }
+            }
+            // Size doesn't exists add results to files
+            let mut temp_hashmap: BTreeMap<String, Vec<FileEntry>> = Default::default();
+            for file_entry in vec_file_entry {
+                temp_hashmap.entry(file_entry.hash.clone()).or_insert_with(Vec::new).push(file_entry);
+            }
+            full_hash_results.push((size, temp_hashmap, Vec::new()));
+        }
+
+        // Must save all results to file, old loaded from file with all currently counted results
+        let mut all_results: BTreeMap<String, FileEntry> = Default::default();
+        for (_size, vec_file_entry) in loaded_hash_map {
+            for file_entry in vec_file_entry {
+                all_results.insert(file_entry.path.to_string_lossy().to_string(), file_entry);
+            }
+        }
+        for (_size, hashmap, _errors) in full_hash_results {
+            for vec_file_entry in hashmap.values() {
+                for file_entry in vec_file_entry {
+                    all_results.insert(file_entry.path.to_string_lossy().to_string(), file_entry.clone());
+                }
+            }
+        }
+        save_hashes_to_file(&all_results, &mut self.text_messages, &self.hash_type, false, self.minimal_cache_file_size);
+    }
+
+    fn full_hashing(
+        &mut self,
+        stop_receiver: Option<&Receiver<()>>,
+        progress_sender: Option<&UnboundedSender<ProgressData>>,
+        pre_checked_map: BTreeMap<u64, Vec<FileEntry>>,
+    ) -> Option<()> {
+        let check_type = self.hash_type;
+
+        let (progress_thread_handle, progress_thread_run, atomic_counter, check_was_stopped) =
+            prepare_thread_handler_common(progress_sender, 2, 2, pre_checked_map.values().map(Vec::len).sum(), self.check_method, self.tool_type);
+
+        ///////////////////////////////////////////////////////////////////////////// HASHING START
+        {
+            let (loaded_hash_map, records_already_cached, non_cached_files_to_check) = self.full_hashing_load_cache_at_start(pre_checked_map);
+
+            let mut full_hash_results: Vec<(u64, BTreeMap<String, Vec<FileEntry>>, Vec<String>)> = non_cached_files_to_check
                 .into_par_iter()
                 .map(|(size, vec_file_entry)| {
                     let mut hashmap_with_hash: BTreeMap<String, Vec<FileEntry>> = Default::default();
                     let mut errors: Vec<String> = Vec::new();
                     let mut buffer = [0u8; 1024 * 16];
 
-                    atomic_file_counter.fetch_add(vec_file_entry.len(), Ordering::Relaxed);
+                    atomic_counter.fetch_add(vec_file_entry.len(), Ordering::Relaxed);
                     for mut file_entry in vec_file_entry {
                         if stop_receiver.is_some() && stop_receiver.unwrap().try_recv().is_ok() {
                             check_was_stopped.store(true, Ordering::Relaxed);
@@ -786,49 +872,13 @@ impl DuplicateFinder {
                 .while_some()
                 .collect();
 
-            if self.use_cache {
-                'main: for (size, vec_file_entry) in records_already_cached {
-                    // Check if size already exists, if exists we must to change it outside because cannot have mut and non mut reference to full_hash_results
-                    for (full_size, full_hashmap, _errors) in &mut full_hash_results {
-                        if size == *full_size {
-                            for file_entry in vec_file_entry {
-                                full_hashmap.entry(file_entry.hash.clone()).or_insert_with(Vec::new).push(file_entry);
-                            }
-                            continue 'main;
-                        }
-                    }
-                    // Size doesn't exists add results to files
-                    let mut temp_hashmap: BTreeMap<String, Vec<FileEntry>> = Default::default();
-                    for file_entry in vec_file_entry {
-                        temp_hashmap.entry(file_entry.hash.clone()).or_insert_with(Vec::new).push(file_entry);
-                    }
-                    full_hash_results.push((size, temp_hashmap, Vec::new()));
-                }
+            self.full_hashing_save_cache_at_exit(records_already_cached, &mut full_hash_results, loaded_hash_map);
 
-                // Must save all results to file, old loaded from file with all currently counted results
-                let mut all_results: BTreeMap<String, FileEntry> = Default::default();
-                for (_size, vec_file_entry) in loaded_hash_map {
-                    for file_entry in vec_file_entry {
-                        all_results.insert(file_entry.path.to_string_lossy().to_string(), file_entry);
-                    }
-                }
-                for (_size, hashmap, _errors) in &full_hash_results {
-                    for vec_file_entry in hashmap.values() {
-                        for file_entry in vec_file_entry {
-                            all_results.insert(file_entry.path.to_string_lossy().to_string(), file_entry.clone());
-                        }
-                    }
-                }
-                save_hashes_to_file(&all_results, &mut self.text_messages, &self.hash_type, false, self.minimal_cache_file_size);
-            }
-
-            // End thread which send info to gui
-            progress_thread_run.store(false, Ordering::Relaxed);
-            progress_thread_handle.join().unwrap();
+            send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
 
             // Break if stop was clicked after saving to cache
             if check_was_stopped.load(Ordering::Relaxed) {
-                return false;
+                return None;
             }
 
             for (size, hash_map, mut errors) in full_hash_results {
@@ -841,27 +891,19 @@ impl DuplicateFinder {
             }
         }
 
-        ///////////////////////////////////////////////////////////////////////////// HASHING END
+        Some(())
+    }
 
+    fn hash_reference_folders(&mut self) {
         // Reference - only use in size, because later hash will be counted differently
         if self.use_reference_folders {
-            let mut btree_map = Default::default();
-            mem::swap(&mut self.files_with_identical_hashes, &mut btree_map);
-            let reference_directories = self.directories.reference_directories.clone();
-            let vec = btree_map
+            let vec = mem::take(&mut self.files_with_identical_hashes)
                 .into_iter()
                 .filter_map(|(_size, vec_vec_file_entry)| {
                     let mut all_results_with_same_size = Vec::new();
                     for vec_file_entry in vec_vec_file_entry {
-                        let mut files_from_referenced_folders = Vec::new();
-                        let mut normal_files = Vec::new();
-                        for file_entry in vec_file_entry {
-                            if reference_directories.iter().any(|e| file_entry.path.starts_with(e)) {
-                                files_from_referenced_folders.push(file_entry);
-                            } else {
-                                normal_files.push(file_entry);
-                            }
-                        }
+                        let (mut files_from_referenced_folders, normal_files): (Vec<_>, Vec<_>) =
+                            vec_file_entry.into_iter().partition(|e| self.directories.is_in_referenced_directory(e.get_path()));
 
                         if files_from_referenced_folders.is_empty() || normal_files.is_empty() {
                             continue;
@@ -897,8 +939,24 @@ impl DuplicateFinder {
                 }
             }
         }
+    }
 
-        Common::print_time(start_time, SystemTime::now(), "check_files_hash - full hash");
+    /// The slowest checking type, which must be applied after checking for size
+    fn check_files_hash(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
+        assert_eq!(self.check_method, CheckingMethod::Hash);
+
+        let mut pre_checked_map: BTreeMap<u64, Vec<FileEntry>> = Default::default();
+        let ret = self.prehashing(stop_receiver, progress_sender, &mut pre_checked_map);
+        if ret.is_none() {
+            return false;
+        }
+
+        let ret = self.full_hashing(stop_receiver, progress_sender, pre_checked_map);
+        if ret.is_none() {
+            return false;
+        }
+
+        self.hash_reference_folders();
 
         // Clean unused data
         self.files_with_identical_size = Default::default();
@@ -909,7 +967,6 @@ impl DuplicateFinder {
     /// Function to delete files, from filed before `BTreeMap`
     /// Using another function to delete files to avoid duplicates data
     fn delete_files(&mut self) {
-        let start_time: SystemTime = SystemTime::now();
         if self.delete_method == DeleteMethod::None {
             return;
         }
@@ -917,6 +974,11 @@ impl DuplicateFinder {
         match self.check_method {
             CheckingMethod::Name => {
                 for vector in self.files_with_identical_names.values() {
+                    let _tuple: (u64, usize, usize) = delete_files(vector, &self.delete_method, &mut self.text_messages, self.dryrun);
+                }
+            }
+            CheckingMethod::SizeName => {
+                for vector in self.files_with_identical_size_names.values() {
                     let _tuple: (u64, usize, usize) = delete_files(vector, &self.delete_method, &mut self.text_messages, self.dryrun);
                 }
             }
@@ -932,13 +994,8 @@ impl DuplicateFinder {
                     let _tuple: (u64, usize, usize) = delete_files(vector, &self.delete_method, &mut self.text_messages, self.dryrun);
                 }
             }
-            CheckingMethod::None => {
-                //Just do nothing
-                panic!("Checking method should never be none.");
-            }
+            _ => panic!(),
         }
-
-        Common::print_time(start_time, SystemTime::now(), "delete_files");
     }
 }
 
@@ -1005,7 +1062,6 @@ impl DebugPrint for DuplicateFinder {
 
 impl SaveResults for DuplicateFinder {
     fn save_results_to_file(&mut self, file_name: &str) -> bool {
-        let start_time: SystemTime = SystemTime::now();
         let file_name: String = match file_name {
             "" => "results.txt".to_string(),
             k => k.to_string(),
@@ -1051,6 +1107,30 @@ impl SaveResults for DuplicateFinder {
                     }
                 } else {
                     write!(writer, "Not found any files with same names.").unwrap();
+                }
+            }
+            CheckingMethod::SizeName => {
+                if !self.files_with_identical_names.is_empty() {
+                    writeln!(
+                        writer,
+                        "-------------------------------------------------Files with same size and names-------------------------------------------------"
+                    )
+                    .unwrap();
+                    writeln!(
+                        writer,
+                        "Found {} files in {} groups with same size and name(may have different content)",
+                        self.information.number_of_duplicated_files_by_size_name, self.information.number_of_groups_by_size_name,
+                    )
+                    .unwrap();
+                    for ((size, name), vector) in self.files_with_identical_size_names.iter().rev() {
+                        writeln!(writer, "Name - {}, {} - {} files ", name, format_size(*size, BINARY), vector.len()).unwrap();
+                        for j in vector {
+                            writeln!(writer, "{}", j.path.display()).unwrap();
+                        }
+                        writeln!(writer).unwrap();
+                    }
+                } else {
+                    write!(writer, "Not found any files with same size and names.").unwrap();
                 }
             }
             CheckingMethod::Size => {
@@ -1105,11 +1185,9 @@ impl SaveResults for DuplicateFinder {
                     write!(writer, "Not found any duplicates.").unwrap();
                 }
             }
-            CheckingMethod::None => {
-                panic!();
-            }
+            _ => panic!(),
         }
-        Common::print_time(start_time, SystemTime::now(), "save_results_to_file");
+
         true
     }
 }
@@ -1118,7 +1196,6 @@ impl PrintResults for DuplicateFinder {
     /// Print information's about duplicated entries
     /// Only needed for CLI
     fn print_results(&self) {
-        let start_time: SystemTime = SystemTime::now();
         let mut number_of_files: u64 = 0;
         let mut number_of_groups: u64 = 0;
 
@@ -1131,6 +1208,20 @@ impl PrintResults for DuplicateFinder {
                 println!("Found {number_of_files} files in {number_of_groups} groups with same name(may have different content)",);
                 for (name, vector) in &self.files_with_identical_names {
                     println!("Name - {} - {} files ", name, vector.len());
+                    for j in vector {
+                        println!("{}", j.path.display());
+                    }
+                    println!();
+                }
+            }
+            CheckingMethod::SizeName => {
+                for i in &self.files_with_identical_size_names {
+                    number_of_files += i.1.len() as u64;
+                    number_of_groups += 1;
+                }
+                println!("Found {number_of_files} files in {number_of_groups} groups with same size and name(may have different content)",);
+                for ((size, name), vector) in &self.files_with_identical_size_names {
+                    println!("Name - {}, {} - {} files ", name, format_size(*size, BINARY), vector.len());
                     for j in vector {
                         println!("{}", j.path.display());
                     }
@@ -1180,11 +1271,8 @@ impl PrintResults for DuplicateFinder {
                     println!();
                 }
             }
-            CheckingMethod::None => {
-                panic!("Checking Method shouldn't be ever set to None");
-            }
+            _ => panic!(),
         }
-        Common::print_time(start_time, SystemTime::now(), "print_entries");
     }
 }
 
@@ -1315,10 +1403,7 @@ pub fn load_hashes_from_file(text_messages: &mut Messages, delete_outdated_cache
         open_cache_folder(&get_file_hash_name(type_of_hash, is_prehash), false, false, &mut text_messages.warnings)
     {
         // Unwrap could fail when failed to open cache file, but json would exists
-        let file_handler = match file_handler {
-            Some(t) => t,
-            _ => return Default::default(),
-        };
+        let Some(file_handler) = file_handler else { return Default::default(); };
         let reader = BufReader::new(file_handler);
 
         let mut hashmap_loaded_entries: BTreeMap<u64, Vec<FileEntry>> = Default::default();
@@ -1443,7 +1528,7 @@ impl MyHasher for crc32fast::Hasher {
     }
 }
 
-impl MyHasher for xxhash_rust::xxh3::Xxh3 {
+impl MyHasher for Xxh3 {
     fn update(&mut self, bytes: &[u8]) {
         self.write(bytes);
     }
