@@ -3,12 +3,11 @@ use std::fs::File;
 use std::io::prelude::*;
 use std::io::BufWriter;
 use std::path::PathBuf;
-use std::time::SystemTime;
 
 use crossbeam_channel::Receiver;
+use futures::channel::mpsc::UnboundedSender;
 
-use crate::common::Common;
-use crate::common_dir_traversal::{Collect, DirTraversalBuilder, DirTraversalResult, ErrorType, FileEntry, ProgressData};
+use crate::common_dir_traversal::{Collect, DirTraversalBuilder, DirTraversalResult, ErrorType, FileEntry, ProgressData, ToolType};
 use crate::common_directory::Directories;
 use crate::common_extensions::Extensions;
 use crate::common_items::ExcludedItems;
@@ -36,6 +35,8 @@ impl Info {
 
 /// Struct with required information's to work
 pub struct InvalidSymlinks {
+    #[allow(dead_code)]
+    tool_type: ToolType,
     text_messages: Messages,
     information: Info,
     invalid_symlinks: Vec<FileEntry>,
@@ -51,6 +52,7 @@ impl InvalidSymlinks {
     #[must_use]
     pub fn new() -> Self {
         Self {
+            tool_type: ToolType::InvalidSymlinks,
             text_messages: Messages::new(),
             information: Info::new(),
             recursive_search: true,
@@ -63,7 +65,7 @@ impl InvalidSymlinks {
         }
     }
 
-    pub fn find_invalid_links(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) {
+    pub fn find_invalid_links(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) {
         self.directories.optimize_directories(self.recursive_search, &mut self.text_messages);
         if !self.check_files(stop_receiver, progress_sender) {
             self.stopped_search = true;
@@ -124,7 +126,7 @@ impl InvalidSymlinks {
     }
 
     /// Check files for any with size == 0
-    fn check_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) -> bool {
+    fn check_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
         let result = DirTraversalBuilder::new()
             .root_dirs(self.directories.included_directories.clone())
             .group_by(|_fe| ())
@@ -138,17 +140,12 @@ impl InvalidSymlinks {
             .build()
             .run();
         match result {
-            DirTraversalResult::SuccessFiles {
-                start_time,
-                grouped_file_entries,
-                warnings,
-            } => {
+            DirTraversalResult::SuccessFiles { grouped_file_entries, warnings } => {
                 if let Some(((), invalid_symlinks)) = grouped_file_entries.into_iter().next() {
                     self.invalid_symlinks = invalid_symlinks;
                 }
                 self.information.number_of_invalid_symlinks = self.invalid_symlinks.len();
                 self.text_messages.warnings.extend(warnings);
-                Common::print_time(start_time, SystemTime::now(), "check_files_name");
                 true
             }
             DirTraversalResult::SuccessFolders { .. } => unreachable!(),
@@ -158,8 +155,6 @@ impl InvalidSymlinks {
 
     /// Function to delete files, from filed Vector
     fn delete_files(&mut self) {
-        let start_time: SystemTime = SystemTime::now();
-
         match self.delete_method {
             DeleteMethod::Delete => {
                 for file_entry in &self.invalid_symlinks {
@@ -172,8 +167,6 @@ impl InvalidSymlinks {
                 //Just do nothing
             }
         }
-
-        Common::print_time(start_time, SystemTime::now(), "delete_files");
     }
 }
 
@@ -215,7 +208,6 @@ impl DebugPrint for InvalidSymlinks {
 
 impl SaveResults for InvalidSymlinks {
     fn save_results_to_file(&mut self, file_name: &str) -> bool {
-        let start_time: SystemTime = SystemTime::now();
         let file_name: String = match file_name {
             "" => "results.txt".to_string(),
             k => k.to_string(),
@@ -257,7 +249,6 @@ impl SaveResults for InvalidSymlinks {
         } else {
             write!(writer, "Not found any invalid symlinks.").unwrap();
         }
-        Common::print_time(start_time, SystemTime::now(), "save_results_to_file");
         true
     }
 }
@@ -266,7 +257,6 @@ impl PrintResults for InvalidSymlinks {
     /// Print information's about duplicated entries
     /// Only needed for CLI
     fn print_results(&self) {
-        let start_time: SystemTime = SystemTime::now();
         println!("Found {} invalid symlinks.\n", self.information.number_of_invalid_symlinks);
         for file_entry in &self.invalid_symlinks {
             println!(
@@ -279,7 +269,5 @@ impl PrintResults for InvalidSymlinks {
                 }
             );
         }
-
-        Common::print_time(start_time, SystemTime::now(), "print_entries");
     }
 }
