@@ -1,17 +1,16 @@
 use std::collections::{BTreeSet, HashMap, HashSet};
-use std::fs::{File, Metadata};
+use std::fs::{DirEntry, File, Metadata};
 use std::io::Write;
 use std::io::*;
+use std::mem;
 use std::panic;
 use std::path::{Path, PathBuf};
-use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-use std::sync::Arc;
-use std::thread::sleep;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-use std::{fs, mem, thread};
+use std::sync::atomic::Ordering;
+use std::time::SystemTime;
 
 use bk_tree::BKTree;
 use crossbeam_channel::Receiver;
+use futures::channel::mpsc::UnboundedSender;
 use humansize::format_size;
 use humansize::BINARY;
 use image::GenericImageView;
@@ -22,16 +21,18 @@ use serde::{Deserialize, Serialize};
 #[cfg(feature = "heif")]
 use crate::common::get_dynamic_image_from_heic;
 use crate::common::{
-    create_crash_message, get_dynamic_image_from_raw_image, get_number_of_threads, open_cache_folder, Common, HEIC_EXTENSIONS, IMAGE_RS_SIMILAR_IMAGES_EXTENSIONS, LOOP_DURATION,
-    RAW_IMAGE_EXTENSIONS,
+    check_folder_children, create_crash_message, get_dynamic_image_from_raw_image, open_cache_folder, prepare_thread_handler_common, send_info_and_wait_for_ending_all_threads,
+    HEIC_EXTENSIONS, IMAGE_RS_SIMILAR_IMAGES_EXTENSIONS, RAW_IMAGE_EXTENSIONS,
 };
+use crate::common_dir_traversal::{common_get_entry_data_metadata, common_read_dir, get_lowercase_name, get_modified_time, CheckingMethod, ProgressData, ToolType};
 use crate::common_directory::Directories;
 use crate::common_extensions::Extensions;
 use crate::common_items::ExcludedItems;
 use crate::common_messages::Messages;
-use crate::common_traits::{DebugPrint, PrintResults, SaveResults};
+use crate::common_traits::{DebugPrint, PrintResults, ResultEntry, SaveResults};
 use crate::flc;
-use crate::localizer_core::generate_translation_hashmap;
+
+type ImHash = Vec<u8>;
 
 pub const SIMILAR_VALUES: [[u32; 6]; 4] = [
     [1, 2, 5, 7, 14, 20],    // 8
@@ -40,22 +41,19 @@ pub const SIMILAR_VALUES: [[u32; 6]; 4] = [
     [6, 20, 40, 40, 40, 40], // 64
 ];
 
-#[derive(Debug)]
-pub struct ProgressData {
-    pub current_stage: u8,
-    pub max_stage: u8,
-    pub images_checked: usize,
-    pub images_to_check: usize,
-}
-
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct FileEntry {
     pub path: PathBuf,
     pub size: u64,
     pub dimensions: String,
     pub modified_date: u64,
-    pub hash: Vec<u8>,
+    pub hash: ImHash,
     pub similarity: u32,
+}
+impl ResultEntry for FileEntry {
+    fn get_path(&self) -> &Path {
+        &self.path
+    }
 }
 
 /// Used by CLI tool when we cannot use directly values
@@ -74,30 +72,32 @@ pub enum SimilarityPreset {
 /// Distance metric to use with the BK-tree.
 struct Hamming;
 
-impl bk_tree::Metric<Vec<u8>> for Hamming {
-    fn distance(&self, a: &Vec<u8>, b: &Vec<u8>) -> u32 {
+impl bk_tree::Metric<ImHash> for Hamming {
+    fn distance(&self, a: &ImHash, b: &ImHash) -> u32 {
         hamming::distance_fast(a, b).unwrap() as u32
     }
 
-    fn threshold_distance(&self, a: &Vec<u8>, b: &Vec<u8>, _threshold: u32) -> Option<u32> {
+    fn threshold_distance(&self, a: &ImHash, b: &ImHash, _threshold: u32) -> Option<u32> {
         Some(self.distance(a, b))
     }
 }
 
 /// Struct to store most basics info about all folder
 pub struct SimilarImages {
+    tool_type: ToolType,
     information: Info,
     text_messages: Messages,
     directories: Directories,
     allowed_extensions: Extensions,
     excluded_items: ExcludedItems,
-    bktree: BKTree<Vec<u8>, Hamming>,
+    bktree: BKTree<ImHash, Hamming>,
     similar_vectors: Vec<Vec<FileEntry>>,
     similar_referenced_vectors: Vec<(FileEntry, Vec<FileEntry>)>,
     recursive_search: bool,
     minimal_file_size: u64,
     maximal_file_size: u64,
-    image_hashes: HashMap<Vec<u8>, Vec<FileEntry>>, // Hashmap with image hashes and Vector with names of files
+    image_hashes: HashMap<ImHash, Vec<FileEntry>>,
+    // Hashmap with image hashes and Vector with names of files
     stopped_search: bool,
     similarity: u32,
     images_to_check: HashMap<String, FileEntry>,
@@ -131,6 +131,7 @@ impl SimilarImages {
     #[must_use]
     pub fn new() -> Self {
         Self {
+            tool_type: ToolType::SimilarImages,
             information: Default::default(),
             text_messages: Messages::new(),
             directories: Directories::new(),
@@ -252,7 +253,7 @@ impl SimilarImages {
     }
 
     /// Public function used by CLI to search for empty folders
-    pub fn find_similar_images(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) {
+    pub fn find_similar_images(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) {
         self.directories.optimize_directories(true, &mut self.text_messages);
         self.use_reference_folders = !self.directories.reference_directories.is_empty();
         if !self.check_for_similar_images(stop_receiver, progress_sender) {
@@ -279,8 +280,7 @@ impl SimilarImages {
 
     /// Function to check if folder are empty.
     /// Parameter `initial_checking` for second check before deleting to be sure that checked folder is still empty
-    fn check_for_similar_images(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) -> bool {
-        let start_time: SystemTime = SystemTime::now();
+    fn check_for_similar_images(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
         let mut folders_to_check: Vec<PathBuf> = Vec::with_capacity(1024 * 2); // This should be small enough too not see to big difference and big enough to store most of paths without needing to resize vector
 
         if !self.allowed_extensions.using_custom_extensions() {
@@ -301,39 +301,12 @@ impl SimilarImages {
             folders_to_check.push(id.clone());
         }
 
-        //// PROGRESS THREAD START
-        let progress_thread_run = Arc::new(AtomicBool::new(true));
-
-        let atomic_file_counter = Arc::new(AtomicUsize::new(0));
-
-        let progress_thread_handle = if let Some(progress_sender) = progress_sender {
-            let progress_send = progress_sender.clone();
-            let progress_thread_run = progress_thread_run.clone();
-            let atomic_file_counter = atomic_file_counter.clone();
-            thread::spawn(move || loop {
-                progress_send
-                    .unbounded_send(ProgressData {
-                        current_stage: 0,
-                        max_stage: 3,
-                        images_checked: atomic_file_counter.load(Ordering::Relaxed),
-                        images_to_check: 0,
-                    })
-                    .unwrap();
-                if !progress_thread_run.load(Ordering::Relaxed) {
-                    break;
-                }
-                sleep(Duration::from_millis(LOOP_DURATION as u64));
-            })
-        } else {
-            thread::spawn(|| {})
-        };
-        //// PROGRESS THREAD END
+        let (progress_thread_handle, progress_thread_run, atomic_counter, _check_was_stopped) =
+            prepare_thread_handler_common(progress_sender, 0, 2, 0, CheckingMethod::None, self.tool_type);
 
         while !folders_to_check.is_empty() {
             if stop_receiver.is_some() && stop_receiver.unwrap().try_recv().is_ok() {
-                // End thread which send info to gui
-                progress_thread_run.store(false, Ordering::Relaxed);
-                progress_thread_handle.join().unwrap();
+                send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
                 return false;
             }
 
@@ -343,120 +316,29 @@ impl SimilarImages {
                     let mut dir_result = vec![];
                     let mut warnings = vec![];
                     let mut fe_result = vec![];
-                    // Read current dir children
-                    let read_dir = match fs::read_dir(current_folder) {
-                        Ok(t) => t,
-                        Err(e) => {
-                            warnings.push(flc!(
-                                "core_cannot_open_dir",
-                                generate_translation_hashmap(vec![("dir", current_folder.display().to_string()), ("reason", e.to_string())])
-                            ));
-                            return (dir_result, warnings, fe_result);
-                        }
+
+                    let Some(read_dir) = common_read_dir(current_folder, &mut warnings) else {
+                        return (dir_result, warnings, fe_result);
                     };
 
-                    // Check every sub folder/file/link etc.
-                    'dir: for entry in read_dir {
-                        let entry_data = match entry {
-                            Ok(t) => t,
-                            Err(e) => {
-                                warnings.push(flc!(
-                                    "core_cannot_read_entry_dir",
-                                    generate_translation_hashmap(vec![("dir", current_folder.display().to_string()), ("reason", e.to_string())])
-                                ));
-                                continue 'dir;
-                            }
+                    for entry in read_dir {
+                        let Some((entry_data, metadata)) = common_get_entry_data_metadata(&entry, &mut warnings, current_folder) else {
+                            continue;
                         };
-                        let metadata: Metadata = match entry_data.metadata() {
-                            Ok(t) => t,
-                            Err(e) => {
-                                warnings.push(flc!(
-                                    "core_cannot_read_metadata_dir",
-                                    generate_translation_hashmap(vec![("dir", current_folder.display().to_string()), ("reason", e.to_string())])
-                                ));
-                                continue 'dir;
-                            }
-                        };
+
                         if metadata.is_dir() {
-                            if !self.recursive_search {
-                                continue 'dir;
-                            }
-
-                            let next_folder = current_folder.join(entry_data.file_name());
-                            if self.directories.is_excluded(&next_folder) {
-                                continue 'dir;
-                            }
-
-                            if self.excluded_items.is_excluded(&next_folder) {
-                                continue 'dir;
-                            }
-
-                            #[cfg(target_family = "unix")]
-                            if self.directories.exclude_other_filesystems() {
-                                match self.directories.is_on_other_filesystems(&next_folder) {
-                                    Ok(true) => continue 'dir,
-                                    Err(e) => warnings.push(e.to_string()),
-                                    _ => (),
-                                }
-                            }
-
-                            dir_result.push(next_folder);
+                            check_folder_children(
+                                &mut dir_result,
+                                &mut warnings,
+                                current_folder,
+                                entry_data,
+                                self.recursive_search,
+                                &self.directories,
+                                &self.excluded_items,
+                            );
                         } else if metadata.is_file() {
-                            atomic_file_counter.fetch_add(1, Ordering::Relaxed);
-
-                            let file_name_lowercase: String = match entry_data.file_name().into_string() {
-                                Ok(t) => t,
-                                Err(_inspected) => {
-                                    warnings.push(flc!(
-                                        "core_file_not_utf8_name",
-                                        generate_translation_hashmap(vec![("name", entry_data.path().display().to_string())])
-                                    ));
-                                    continue 'dir;
-                                }
-                            }
-                            .to_lowercase();
-
-                            if !self.allowed_extensions.matches_filename(&file_name_lowercase) {
-                                continue 'dir;
-                            }
-
-                            // Checking files
-                            if (self.minimal_file_size..=self.maximal_file_size).contains(&metadata.len()) {
-                                let current_file_name = current_folder.join(entry_data.file_name());
-                                if self.excluded_items.is_excluded(&current_file_name) {
-                                    continue 'dir;
-                                }
-
-                                let fe: FileEntry = FileEntry {
-                                    path: current_file_name.clone(),
-                                    size: metadata.len(),
-                                    dimensions: String::new(),
-                                    modified_date: match metadata.modified() {
-                                        Ok(t) => match t.duration_since(UNIX_EPOCH) {
-                                            Ok(d) => d.as_secs(),
-                                            Err(_inspected) => {
-                                                warnings.push(flc!(
-                                                    "core_file_modified_before_epoch",
-                                                    generate_translation_hashmap(vec![("name", current_file_name.display().to_string())])
-                                                ));
-                                                0
-                                            }
-                                        },
-                                        Err(e) => {
-                                            warnings.push(flc!(
-                                                "core_file_no_modification_date",
-                                                generate_translation_hashmap(vec![("name", current_file_name.display().to_string()), ("reason", e.to_string())])
-                                            ));
-                                            0
-                                        }
-                                    },
-
-                                    hash: Vec::new(),
-                                    similarity: 0,
-                                };
-
-                                fe_result.push((current_file_name.to_string_lossy().to_string(), fe));
-                            }
+                            atomic_counter.fetch_add(1, Ordering::Relaxed);
+                            self.add_file_entry(&metadata, current_folder, entry_data, &mut fe_result, &mut warnings);
                         }
                     }
                     (dir_result, warnings, fe_result)
@@ -476,23 +358,41 @@ impl SimilarImages {
             }
         }
 
-        // End thread which send info to gui
-        progress_thread_run.store(false, Ordering::Relaxed);
-        progress_thread_handle.join().unwrap();
-        Common::print_time(start_time, SystemTime::now(), "check_for_similar_images");
+        send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
+
         true
     }
 
-    // Cache algorithm:
-    // - Load data from file
-    // - Remove from data to search, already loaded entries from cache(size and modified datamust match)
-    // - Check hash of files which doesn't have saved entry
-    // - Join already read hashes with hashes which were read from file
-    // - Join all hashes and save it to file
+    fn add_file_entry(&self, metadata: &Metadata, current_folder: &Path, entry_data: &DirEntry, fe_result: &mut Vec<(String, FileEntry)>, warnings: &mut Vec<String>) {
+        let Some(file_name_lowercase) = get_lowercase_name(entry_data, warnings) else {
+            return;
+        };
 
-    fn hash_images(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) -> bool {
-        let hash_map_modification = SystemTime::now();
+        if !self.allowed_extensions.matches_filename(&file_name_lowercase) {
+            return;
+        }
 
+        // Checking files
+        if (self.minimal_file_size..=self.maximal_file_size).contains(&metadata.len()) {
+            let current_file_name = current_folder.join(entry_data.file_name());
+            if self.excluded_items.is_excluded(&current_file_name) {
+                return;
+            }
+
+            let fe: FileEntry = FileEntry {
+                path: current_file_name.clone(),
+                size: metadata.len(),
+                dimensions: String::new(),
+                modified_date: get_modified_time(metadata, warnings, &current_file_name, false),
+                hash: Vec::new(),
+                similarity: 0,
+            };
+
+            fe_result.push((current_file_name.to_string_lossy().to_string(), fe));
+        }
+    }
+
+    fn hash_images_load_cache(&mut self) -> (HashMap<String, FileEntry>, HashMap<String, FileEntry>, HashMap<String, FileEntry>) {
         let loaded_hash_map;
 
         let mut records_already_cached: HashMap<String, FileEntry> = Default::default();
@@ -505,147 +405,59 @@ impl SimilarImages {
             };
 
             for (name, file_entry) in &self.images_to_check {
-                #[allow(clippy::if_same_then_else)]
                 if !loaded_hash_map.contains_key(name) {
                     // If loaded data doesn't contains current image info
                     non_cached_files_to_check.insert(name.clone(), file_entry.clone());
-                } else if file_entry.size != loaded_hash_map.get(name).unwrap().size || file_entry.modified_date != loaded_hash_map.get(name).unwrap().modified_date {
-                    // When size or modification date of image changed, then it is clear that is different image
-                    non_cached_files_to_check.insert(name.clone(), file_entry.clone());
                 } else {
-                    // Checking may be omitted when already there is entry with same size and modification date
-                    records_already_cached.insert(name.clone(), loaded_hash_map.get(name).unwrap().clone());
+                    let loaded_item = loaded_hash_map.get(name).unwrap();
+                    if file_entry.size != loaded_item.size || file_entry.modified_date != loaded_item.modified_date {
+                        // When size or modification date of image changed, then it is clear that is different image
+                        non_cached_files_to_check.insert(name.clone(), file_entry.clone());
+                    } else {
+                        // Checking may be omitted when already there is entry with same size and modification date
+                        records_already_cached.insert(name.clone(), loaded_item.clone());
+                    }
                 }
             }
         } else {
             loaded_hash_map = Default::default();
             mem::swap(&mut self.images_to_check, &mut non_cached_files_to_check);
         }
+        (loaded_hash_map, records_already_cached, non_cached_files_to_check)
+    }
 
-        Common::print_time(hash_map_modification, SystemTime::now(), "sort_images - reading data from cache and preparing them");
-        let hash_map_modification = SystemTime::now();
+    // Cache algorithm:
+    // - Load data from file
+    // - Remove from data to search, already loaded entries from cache(size and modified date must match)
+    // - Check hash of files which doesn't have saved entry
+    // - Join already read hashes with hashes which were read from file
+    // - Join all hashes and save it to file
 
-        //// PROGRESS THREAD START
-        let check_was_stopped = AtomicBool::new(false); // Used for breaking from GUI and ending check thread
-        let progress_thread_run = Arc::new(AtomicBool::new(true));
+    fn hash_images(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
+        let (loaded_hash_map, records_already_cached, non_cached_files_to_check) = self.hash_images_load_cache();
 
-        let atomic_file_counter = Arc::new(AtomicUsize::new(0));
+        let (progress_thread_handle, progress_thread_run, atomic_counter, check_was_stopped) =
+            prepare_thread_handler_common(progress_sender, 1, 2, non_cached_files_to_check.len(), CheckingMethod::None, self.tool_type);
 
-        let progress_thread_handle = if let Some(progress_sender) = progress_sender {
-            let progress_send = progress_sender.clone();
-            let progress_thread_run = progress_thread_run.clone();
-            let atomic_file_counter = atomic_file_counter.clone();
-            let images_to_check = non_cached_files_to_check.len();
-            thread::spawn(move || loop {
-                progress_send
-                    .unbounded_send(ProgressData {
-                        current_stage: 1,
-                        max_stage: 3,
-                        images_checked: atomic_file_counter.load(Ordering::Relaxed),
-                        images_to_check,
-                    })
-                    .unwrap();
-                if !progress_thread_run.load(Ordering::Relaxed) {
-                    break;
-                }
-                sleep(Duration::from_millis(LOOP_DURATION as u64));
-            })
-        } else {
-            thread::spawn(|| {})
-        };
-
-        //// PROGRESS THREAD END
-        let mut vec_file_entry: Vec<(FileEntry, Vec<u8>)> = non_cached_files_to_check
+        let mut vec_file_entry: Vec<(FileEntry, ImHash)> = non_cached_files_to_check
             .into_par_iter()
-            .map(|(_s, mut file_entry)| {
-                atomic_file_counter.fetch_add(1, Ordering::Relaxed);
+            .map(|(_s, file_entry)| {
+                atomic_counter.fetch_add(1, Ordering::Relaxed);
                 if stop_receiver.is_some() && stop_receiver.unwrap().try_recv().is_ok() {
                     check_was_stopped.store(true, Ordering::Relaxed);
                     return None;
                 }
-                let file_name_lowercase = file_entry.path.to_string_lossy().to_lowercase();
-
-                let image;
-
-                #[allow(clippy::never_loop)] // Required to implement nice if/else
-                'krztyna: loop {
-                    if RAW_IMAGE_EXTENSIONS.iter().any(|e| file_name_lowercase.ends_with(e)) {
-                        image = match get_dynamic_image_from_raw_image(&file_entry.path) {
-                            Some(t) => t,
-                            None => return Some(Some((file_entry, Vec::new()))),
-                        };
-                        break 'krztyna;
-                    }
-
-                    #[cfg(feature = "heif")]
-                    if HEIC_EXTENSIONS.iter().any(|e| file_name_lowercase.ends_with(e)) {
-                        image = match get_dynamic_image_from_heic(&file_entry.path.to_string_lossy()) {
-                            Ok(t) => t,
-                            Err(_) => {
-                                return Some(Some((file_entry, Vec::new())));
-                            }
-                        };
-                        break 'krztyna;
-                    }
-
-                    // Normal image extension, when any other fail, not using if/else
-                    let result = panic::catch_unwind(|| {
-                        match image::open(file_entry.path.clone()) {
-                            Ok(t) => Ok(t),
-                            // Err(_inspected) => return Some(None), // Something is wrong with image,
-                            // For broken images empty hash is used, because without it will try to resecan files each time when it is called(missing cache file is responsible for it)
-                            // This may cause problems(very rarely), when e.g. file was not available due lack of permissions, but it is available now
-                            Err(_inspected) => Err(()),
-                        }
-                    });
-
-                    // If image crashed during opening, we just skip checking its hash and go on
-                    if let Ok(image_result) = result {
-                        if let Ok(image2) = image_result {
-                            image = image2;
-                        } else {
-                            return Some(Some((file_entry, Vec::new())));
-                        }
-                    } else {
-                        let message = create_crash_message("Image-rs", &file_entry.path.to_string_lossy(), "https://github.com/image-rs/image/issues");
-                        println!("{message}");
-                        return Some(Some((file_entry, Vec::new())));
-                    }
-
-                    break 'krztyna;
-                }
-
-                let dimensions = image.dimensions();
-
-                file_entry.dimensions = format!("{}x{}", dimensions.0, dimensions.1);
-
-                let hasher_config = HasherConfig::new()
-                    .hash_size(self.hash_size as u32, self.hash_size as u32)
-                    .hash_alg(self.hash_alg)
-                    .resize_filter(self.image_filter);
-                let hasher = hasher_config.to_hasher();
-
-                let hash = hasher.hash_image(&image);
-                let buf: Vec<u8> = hash.as_bytes().to_vec();
-
-                file_entry.hash = buf.clone();
-
-                Some(Some((file_entry, buf)))
+                Some(Some(self.collect_image_file_entry(file_entry)))
             })
             .while_some()
             .filter(Option::is_some)
             .map(Option::unwrap)
-            .collect::<Vec<(FileEntry, Vec<u8>)>>();
+            .collect::<Vec<(FileEntry, ImHash)>>();
 
-        // End thread which send info to gui
-        progress_thread_run.store(false, Ordering::Relaxed);
-        progress_thread_handle.join().unwrap();
-
-        Common::print_time(hash_map_modification, SystemTime::now(), "sort_images - reading data from files in parallel");
-        let hash_map_modification = SystemTime::now();
+        send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
 
         // Just connect loaded results with already calculated hashes
-        for (_name, file_entry) in records_already_cached {
+        for file_entry in records_already_cached.into_values() {
             vec_file_entry.push((file_entry.clone(), file_entry.hash));
         }
 
@@ -678,361 +490,349 @@ impl SimilarImages {
             return false;
         }
 
-        Common::print_time(hash_map_modification, SystemTime::now(), "sort_images - saving data to files");
+        true
+    }
+    fn collect_image_file_entry(&self, mut file_entry: FileEntry) -> (FileEntry, ImHash) {
+        let file_name_lowercase = file_entry.path.to_string_lossy().to_lowercase();
+
+        let image;
+
+        #[allow(clippy::never_loop)] // Required to implement nice if/else
+        'krztyna: loop {
+            if RAW_IMAGE_EXTENSIONS.iter().any(|e| file_name_lowercase.ends_with(e)) {
+                image = match get_dynamic_image_from_raw_image(&file_entry.path) {
+                    Some(t) => t,
+                    None => return (file_entry, Vec::new()),
+                };
+                break 'krztyna;
+            }
+
+            #[cfg(feature = "heif")]
+            if HEIC_EXTENSIONS.iter().any(|e| file_name_lowercase.ends_with(e)) {
+                image = match get_dynamic_image_from_heic(&file_entry.path.to_string_lossy()) {
+                    Ok(t) => t,
+                    Err(_) => {
+                        return (file_entry, Vec::new());
+                    }
+                };
+                break 'krztyna;
+            }
+
+            // Normal image extension, when any other fail, not using if/else
+            let result = panic::catch_unwind(|| {
+                match image::open(file_entry.path.clone()) {
+                    Ok(t) => Ok(t),
+                    // Err(_inspected) => return Some(None), // Something is wrong with image,
+                    // For broken images empty hash is used, because without it will try to resecan files each time when it is called(missing cache file is responsible for it)
+                    // This may cause problems(very rarely), when e.g. file was not available due lack of permissions, but it is available now
+                    Err(_inspected) => Err(()),
+                }
+            });
+
+            // If image crashed during opening, we just skip checking its hash and go on
+            if let Ok(image_result) = result {
+                if let Ok(image2) = image_result {
+                    image = image2;
+                } else {
+                    return (file_entry, Vec::new());
+                }
+            } else {
+                let message = create_crash_message("Image-rs", &file_entry.path.to_string_lossy(), "https://github.com/image-rs/image/issues");
+                println!("{message}");
+                return (file_entry, Vec::new());
+            }
+
+            break 'krztyna;
+        }
+
+        let dimensions = image.dimensions();
+
+        file_entry.dimensions = format!("{}x{}", dimensions.0, dimensions.1);
+
+        let hasher_config = HasherConfig::new()
+            .hash_size(self.hash_size as u32, self.hash_size as u32)
+            .hash_alg(self.hash_alg)
+            .resize_filter(self.image_filter);
+        let hasher = hasher_config.to_hasher();
+
+        let hash = hasher.hash_image(&image);
+        let buf: ImHash = hash.as_bytes().to_vec();
+
+        file_entry.hash = buf.clone();
+
+        (file_entry, buf)
+    }
+
+    // Split hashes at 2 parts, base hashes and hashes to compare, 3 argument is set of hashes with multiple images
+    fn split_hashes(&mut self, all_hashed_images: &HashMap<ImHash, Vec<FileEntry>>) -> (Vec<ImHash>, HashSet<ImHash>) {
+        let hashes_with_multiple_images: HashSet<ImHash> = all_hashed_images
+            .iter()
+            .filter_map(|(hash, vec_file_entry)| {
+                if vec_file_entry.len() >= 2 {
+                    return Some(hash.clone());
+                };
+                None
+            })
+            .collect();
+        let mut base_hashes = Vec::new(); // Initial hashes
+        if self.use_reference_folders {
+            let mut files_from_referenced_folders: HashMap<ImHash, Vec<FileEntry>> = HashMap::new();
+            let mut normal_files: HashMap<ImHash, Vec<FileEntry>> = HashMap::new();
+
+            all_hashed_images.clone().into_iter().for_each(|(hash, vec_file_entry)| {
+                for file_entry in vec_file_entry {
+                    if is_in_reference_folder(&self.directories.reference_directories, &file_entry.path) {
+                        files_from_referenced_folders.entry(hash.clone()).or_insert_with(Vec::new).push(file_entry);
+                    } else {
+                        normal_files.entry(hash.clone()).or_insert_with(Vec::new).push(file_entry);
+                    }
+                }
+            });
+
+            for hash in normal_files.into_keys() {
+                self.bktree.add(hash);
+            }
+
+            for hash in files_from_referenced_folders.into_keys() {
+                base_hashes.push(hash);
+            }
+        } else {
+            for original_hash in all_hashed_images.keys() {
+                self.bktree.add(original_hash.clone());
+            }
+            base_hashes = all_hashed_images.keys().cloned().collect::<Vec<_>>();
+        }
+        (base_hashes, hashes_with_multiple_images)
+    }
+
+    fn collect_hash_compare_result(
+        &self,
+        hashes_parents: HashMap<ImHash, u32>,
+        hashes_with_multiple_images: &HashSet<ImHash>,
+        all_hashed_images: &HashMap<ImHash, Vec<FileEntry>>,
+        collected_similar_images: &mut HashMap<ImHash, Vec<FileEntry>>,
+        hashes_similarity: HashMap<ImHash, (ImHash, u32)>,
+    ) {
+        if self.use_reference_folders {
+            // This is same step as without reference folders, but also checks if children are inside/outside reference directories, because may happen, that one file is inside reference folder and other outside
+
+            // Collecting results to vector
+            for (parent_hash, child_number) in hashes_parents {
+                // If hash contains other hasher OR multiple images are available for checked hash
+                if child_number > 0 || hashes_with_multiple_images.contains(&parent_hash) {
+                    let vec_fe = all_hashed_images
+                        .get(&parent_hash)
+                        .unwrap()
+                        .iter()
+                        .filter(|e| is_in_reference_folder(&self.directories.reference_directories, &e.path))
+                        .cloned()
+                        .collect();
+                    collected_similar_images.insert(parent_hash.clone(), vec_fe);
+                }
+            }
+
+            for (child_hash, (parent_hash, similarity)) in hashes_similarity {
+                let mut vec_fe: Vec<_> = all_hashed_images
+                    .get(&child_hash)
+                    .unwrap()
+                    .iter()
+                    .filter(|e| !is_in_reference_folder(&self.directories.reference_directories, &e.path))
+                    .cloned()
+                    .collect();
+                for fe in &mut vec_fe {
+                    fe.similarity = similarity;
+                }
+                collected_similar_images.get_mut(&parent_hash).unwrap().append(&mut vec_fe);
+            }
+        } else {
+            // Collecting results to vector
+            for (parent_hash, child_number) in hashes_parents {
+                // If hash contains other hasher OR multiple images are available for checked hash
+                if child_number > 0 || hashes_with_multiple_images.contains(&parent_hash) {
+                    let vec_fe = all_hashed_images.get(&parent_hash).unwrap().clone();
+                    collected_similar_images.insert(parent_hash.clone(), vec_fe);
+                }
+            }
+
+            for (child_hash, (parent_hash, similarity)) in hashes_similarity {
+                let mut vec_fe = all_hashed_images.get(&child_hash).unwrap().clone();
+                for fe in &mut vec_fe {
+                    fe.similarity = similarity;
+                }
+                collected_similar_images.get_mut(&parent_hash).unwrap().append(&mut vec_fe);
+            }
+        }
+    }
+
+    fn compare_hashes_with_non_zero_tolerance(
+        &mut self,
+        all_hashed_images: &HashMap<ImHash, Vec<FileEntry>>,
+        collected_similar_images: &mut HashMap<ImHash, Vec<FileEntry>>,
+        progress_sender: Option<&UnboundedSender<ProgressData>>,
+        stop_receiver: Option<&Receiver<()>>,
+        tolerance: u32,
+    ) -> bool {
+        // Don't use hashes with multiple images in bktree, because they will always be master of group and cannot be find by other hashes
+        let (base_hashes, hashes_with_multiple_images) = self.split_hashes(all_hashed_images);
+
+        let (progress_thread_handle, progress_thread_run, atomic_counter, check_was_stopped) =
+            prepare_thread_handler_common(progress_sender, 2, 2, base_hashes.len(), CheckingMethod::None, self.tool_type);
+
+        let mut hashes_parents: HashMap<ImHash, u32> = Default::default(); // Hashes used as parent (hash, children_number_of_hash)
+        let mut hashes_similarity: HashMap<ImHash, (ImHash, u32)> = Default::default(); // Hashes used as child, (parent_hash, similarity)
+
+        // Check them in chunks, to decrease number of used memory
+        // println!();
+        let base_hashes_chunks = base_hashes.chunks(1000);
+        for chunk in base_hashes_chunks {
+            let partial_results = chunk
+                .into_par_iter()
+                .map(|hash_to_check| {
+                    atomic_counter.fetch_add(1, Ordering::Relaxed);
+
+                    if stop_receiver.is_some() && stop_receiver.unwrap().try_recv().is_ok() {
+                        check_was_stopped.store(true, Ordering::Relaxed);
+                        return None;
+                    }
+                    let mut found_items = self
+                        .bktree
+                        .find(hash_to_check, tolerance)
+                        .filter(|(similarity, compared_hash)| {
+                            *similarity != 0 && !hashes_parents.contains_key(*compared_hash) && !hashes_with_multiple_images.contains(*compared_hash)
+                        })
+                        .filter(|(similarity, compared_hash)| {
+                            if let Some((_, other_similarity_with_parent)) = hashes_similarity.get(*compared_hash) {
+                                // If current hash is more similar to other hash than to current parent hash, then skip check earlier
+                                // Because there is no way to be more similar to other hash than to current parent hash
+                                if *similarity >= *other_similarity_with_parent {
+                                    return false;
+                                }
+                            }
+                            true
+                        })
+                        .collect::<Vec<_>>();
+
+                    found_items.sort_unstable_by_key(|f| f.0);
+                    Some((hash_to_check, found_items))
+                })
+                .while_some()
+                .filter(|(original_hash, vec_similar_hashes)| !vec_similar_hashes.is_empty() || hashes_with_multiple_images.contains(*original_hash))
+                .collect::<Vec<_>>();
+
+            // for (hash, vec) in &partial_results {
+            //     println!("{hash:?} --- {:?}", vec.iter().map(|e| e.1).collect::<Vec<_>>());
+            // }
+
+            if check_was_stopped.load(Ordering::Relaxed) {
+                send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
+                return false;
+            }
+
+            self.connect_results(partial_results, &mut hashes_parents, &mut hashes_similarity, &hashes_with_multiple_images);
+        }
+
+        send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
+
+        debug_check_for_duplicated_things(self.use_reference_folders, &hashes_parents, &hashes_similarity, all_hashed_images, "LATTER");
+        self.collect_hash_compare_result(hashes_parents, &hashes_with_multiple_images, all_hashed_images, collected_similar_images, hashes_similarity);
+
         true
     }
 
-    fn find_similar_hashes(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&futures::channel::mpsc::UnboundedSender<ProgressData>>) -> bool {
+    fn connect_results(
+        &self,
+        partial_results: Vec<(&ImHash, Vec<(u32, &ImHash)>)>,
+        hashes_parents: &mut HashMap<ImHash, u32>,
+        hashes_similarity: &mut HashMap<ImHash, (ImHash, u32)>,
+        hashes_with_multiple_images: &HashSet<ImHash>,
+    ) {
+        for (original_hash, vec_compared_hashes) in partial_results {
+            let mut number_of_added_child_items = 0;
+            for (similarity, compared_hash) in vec_compared_hashes {
+                // If hash is already in results skip it
+                // This check duplicates check from bktree.find, but it is needed to because when iterating over elements, this structure can change
+                if hashes_parents.contains_key(compared_hash) {
+                    continue;
+                }
+
+                // If there is already record, with smaller sensitivity, then replace it
+                let mut need_to_add = false;
+                let mut need_to_check = false;
+
+                // TODO consider to replace variables from above with closures
+                // If current checked hash, have parent, first we must check if similarity between them is lower than checked item
+                if let Some((current_parent_hash, current_similarity_with_parent)) = hashes_similarity.get(original_hash) {
+                    if *current_similarity_with_parent > similarity {
+                        need_to_check = true;
+
+                        *hashes_parents.get_mut(current_parent_hash).unwrap() -= 1;
+                        if hashes_parents.get(current_parent_hash) == Some(&0) && !hashes_with_multiple_images.contains(current_parent_hash) {
+                            hashes_parents.remove(current_parent_hash);
+                        }
+                        hashes_similarity.remove(original_hash).unwrap();
+                    }
+                } else {
+                    need_to_check = true;
+                }
+
+                if need_to_check {
+                    if let Some((other_parent_hash, other_similarity)) = hashes_similarity.get(compared_hash) {
+                        if *other_similarity > similarity {
+                            need_to_add = true;
+                            *hashes_parents.get_mut(other_parent_hash).unwrap() -= 1;
+                        }
+                    }
+                    // But when there is no record, just add it
+                    else {
+                        need_to_add = true;
+                    }
+                }
+
+                if need_to_add {
+                    hashes_similarity.insert(compared_hash.clone(), (original_hash.clone(), similarity));
+                    number_of_added_child_items += 1;
+                }
+            }
+
+            if number_of_added_child_items > 0 || hashes_with_multiple_images.contains(original_hash) {
+                hashes_parents.insert((*original_hash).clone(), number_of_added_child_items);
+            }
+        }
+    }
+
+    fn find_similar_hashes(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
         if self.image_hashes.is_empty() {
             return true;
         }
 
-        let hash_map_modification = SystemTime::now();
         let tolerance = self.similarity;
 
         // Results
-        let mut collected_similar_images: HashMap<Vec<u8>, Vec<FileEntry>> = Default::default();
+        let mut collected_similar_images: HashMap<ImHash, Vec<FileEntry>> = Default::default();
 
-        let mut all_hashed_images = Default::default();
-        mem::swap(&mut all_hashed_images, &mut self.image_hashes);
-
-        let all_hashes: Vec<_> = all_hashed_images.keys().collect();
+        let all_hashed_images = mem::take(&mut self.image_hashes);
 
         // Checking entries with tolerance 0 is really easy and fast, because only entries with same hashes needs to be checked
         if tolerance == 0 {
-            for (hash, vec_file_entry) in all_hashed_images.clone() {
+            for (hash, vec_file_entry) in all_hashed_images {
                 if vec_file_entry.len() >= 2 {
                     collected_similar_images.insert(hash, vec_file_entry);
                 }
             }
         } else {
-            //// PROGRESS THREAD START
-            let check_was_stopped = AtomicBool::new(false); // Used for breaking from GUI and ending check thread
-            let progress_thread_run = Arc::new(AtomicBool::new(true));
-            let atomic_mode_counter = Arc::new(AtomicUsize::new(0));
-
-            let progress_thread_handle = if let Some(progress_sender) = progress_sender {
-                let progress_send = progress_sender.clone();
-                let progress_thread_run = progress_thread_run.clone();
-                let atomic_mode_counter = atomic_mode_counter.clone();
-                let all_combinations_to_check = all_hashes.len();
-                thread::spawn(move || loop {
-                    progress_send
-                        .unbounded_send(ProgressData {
-                            current_stage: 2,
-                            max_stage: 2,
-                            images_checked: atomic_mode_counter.load(Ordering::Relaxed),
-                            images_to_check: all_combinations_to_check,
-                        })
-                        .unwrap();
-                    if !progress_thread_run.load(Ordering::Relaxed) {
-                        break;
-                    }
-                    sleep(Duration::from_millis(LOOP_DURATION as u64));
-                })
-            } else {
-                thread::spawn(|| {})
-            };
-            //// PROGRESS THREAD END
-
-            // Don't use hashes with multiple images in bktree, because they will always be master of group and cannot be find by other hashes
-            let mut hashes_with_multiple_images: HashSet<_> = Default::default(); // Fast way to check if hash have multiple images
-
-            let mut files_from_referenced_folders = HashMap::new();
-            let mut normal_files = HashMap::new();
-
-            let number_of_processors = get_number_of_threads();
-            let chunk_size;
-            let mut chunks: Vec<&[&Vec<u8>]>;
-
-            let mut initial_hashes: Vec<&Vec<u8>> = Vec::new();
-            let mut additional_chunk_to_check: Vec<&Vec<u8>> = Default::default();
-
-            if self.use_reference_folders {
-                let reference_directories = self.directories.reference_directories.clone();
-                all_hashed_images.clone().into_iter().for_each(|(hash, vec_file_entry)| {
-                    for file_entry in vec_file_entry {
-                        if reference_directories.iter().any(|e| file_entry.path.starts_with(e)) {
-                            files_from_referenced_folders.entry(hash.clone()).or_insert_with(Vec::new).push(file_entry);
-                        } else {
-                            normal_files.entry(hash.clone()).or_insert_with(Vec::new).push(file_entry);
-                        }
-                    }
-                });
-
-                for (hash, vec_files) in &normal_files {
-                    if vec_files.len() >= 2 {
-                        hashes_with_multiple_images.insert(hash);
-                    }
-                    self.bktree.add(hash.clone());
-                }
-                for (hash, vec_files) in &files_from_referenced_folders {
-                    if vec_files.len() >= 2 {
-                        hashes_with_multiple_images.insert(hash);
-                    }
-                    initial_hashes.push(hash);
-                }
-                chunk_size = initial_hashes.len() / number_of_processors;
-
-                chunks = if chunk_size > 0 {
-                    initial_hashes.chunks(chunk_size).collect::<Vec<_>>()
-                } else {
-                    vec![&initial_hashes]
-                };
-            } else {
-                for (hash, vec_files) in &all_hashed_images {
-                    if vec_files.len() >= 2 {
-                        additional_chunk_to_check.push(hash);
-                        hashes_with_multiple_images.insert(hash);
-                    } else {
-                        self.bktree.add(hash.clone());
-                    }
-                }
-                chunk_size = all_hashes.len() / number_of_processors;
-                chunks = if chunk_size > 0 {
-                    all_hashes.chunks(chunk_size).collect::<Vec<_>>()
-                } else {
-                    vec![&all_hashes]
-                };
-                chunks.push(&additional_chunk_to_check);
-            }
-
-            let parts: Vec<_> = chunks
-                .into_par_iter()
-                .map(|hashes_to_check| {
-                    let mut hashes_parents: HashMap<&Vec<u8>, u32> = Default::default(); // Hashes used as parent (hash, children_number_of_hash)
-                    let mut hashes_similarity: HashMap<&Vec<u8>, (&Vec<u8>, u32)> = Default::default(); // Hashes used as child, (parent_hash, similarity)
-
-                    // Sprawdź czy hash nie jest użyty jako master gdzie indziej
-                    // Jeśli tak to przejdź do sprawdzania kolejnego elementu
-                    // Zweryfikuj czy sprawdzany element ma rodzica
-                    // Jeśli ma to sprawdź czy similarity nowego rodzica jest mniejsze niż starego
-                    // // Jeśli tak to zmniejsz ilość dzieci starego rodzica, dodaj ilość dzieci w nowym rodzicu i podmień rekord hashes_similarity
-                    // // Jeśli nie to dodaj nowy rekord w hashes_similarity jak i hashes_parents z liczbą dzieci równą 1
-
-                    for (index, hash_to_check) in hashes_to_check.iter().enumerate() {
-                        // Don't check for user stop too often
-                        // Also don't add too often data to atomic variable
-                        const CYCLES_COUNTER: usize = 0b11_1111;
-                        if ((index & CYCLES_COUNTER) == CYCLES_COUNTER) && index != 0 {
-                            atomic_mode_counter.fetch_add(CYCLES_COUNTER, Ordering::Relaxed);
-                            if stop_receiver.is_some() && stop_receiver.unwrap().try_recv().is_ok() {
-                                check_was_stopped.store(true, Ordering::Relaxed);
-                                return None;
-                            }
-                        }
-                        hashes_parents.insert(hash_to_check, 0);
-
-                        let mut found_items = self
-                            .bktree
-                            .find(hash_to_check, tolerance)
-                            .filter(|(similarity, _hash)| if self.use_reference_folders { true } else { *similarity != 0 })
-                            .collect::<Vec<_>>();
-
-                        found_items.sort_unstable_by_key(|f| f.0);
-
-                        for (similarity, compared_hash) in found_items {
-                            image_to_check(
-                                &mut hashes_parents,
-                                &mut hashes_similarity,
-                                &hashes_with_multiple_images,
-                                hash_to_check,
-                                compared_hash,
-                                similarity,
-                            );
-                        }
-                    }
-
-                    #[cfg(debug_assertions)]
-                    if !self.use_reference_folders {
-                        debug_check_for_duplicated_things(&hashes_parents, &hashes_similarity, &all_hashed_images, "BEFORE");
-                    }
-
-                    Some((hashes_parents, hashes_similarity))
-                })
-                .while_some()
-                .collect();
-
-            // End thread which send info to gui
-            progress_thread_run.store(false, Ordering::Relaxed);
-            progress_thread_handle.join().unwrap();
-
-            if check_was_stopped.load(Ordering::Relaxed) {
+            if !self.compare_hashes_with_non_zero_tolerance(&all_hashed_images, &mut collected_similar_images, progress_sender, stop_receiver, tolerance) {
                 return false;
             }
-
-            {
-                let mut hashes_parents: HashMap<&Vec<u8>, u32> = Default::default();
-                let mut hashes_similarity: HashMap<&Vec<u8>, (&Vec<u8>, u32)> = Default::default();
-                let mut iter = parts.into_iter();
-                // At start fill arrays with first item
-                // Normal algorithm would do exactly same thing, but slower, one record after one
-                if let Some((first_hashes_parents, first_hashes_similarity)) = iter.next() {
-                    hashes_parents = first_hashes_parents;
-                    hashes_similarity = first_hashes_similarity;
-                }
-
-                for (partial_hashes_with_parents, partial_hashes_with_similarity) in iter {
-                    for (parent_hash, _child_number) in partial_hashes_with_parents {
-                        if !hashes_parents.contains_key(parent_hash) && !hashes_similarity.contains_key(parent_hash) {
-                            hashes_parents.insert(parent_hash, 0);
-                        }
-                    }
-
-                    for (hash_to_check, (compared_hash, similarity)) in partial_hashes_with_similarity {
-                        image_to_check(
-                            &mut hashes_parents,
-                            &mut hashes_similarity,
-                            &hashes_with_multiple_images,
-                            hash_to_check,
-                            compared_hash,
-                            similarity,
-                        );
-                    }
-                }
-
-                #[cfg(debug_assertions)]
-                if !self.use_reference_folders {
-                    debug_check_for_duplicated_things(&hashes_parents, &hashes_similarity, &all_hashed_images, "LATTER");
-                }
-
-                // Just simple check if all original hashes with multiple entries are available in end results
-                let original_hashes_at_start = hashes_with_multiple_images.len();
-                let original_hashes_in_end_results = hashes_parents
-                    .iter()
-                    .filter(|(parent_hash, _child_number)| hashes_with_multiple_images.contains(*parent_hash))
-                    .count();
-                if !self.use_reference_folders {
-                    assert_eq!(original_hashes_at_start, original_hashes_in_end_results);
-                }
-
-                if self.use_reference_folders {
-                    // This is same step as without reference folders, but also checks if children are inside/outside reference directories, because may happen, that one file is inside reference folder and other outside
-
-                    // Collecting results to vector
-                    for (parent_hash, child_number) in hashes_parents {
-                        // If hash contains other hasher OR multiple images are available for checked hash
-                        if child_number > 0 || hashes_with_multiple_images.contains(parent_hash) {
-                            let vec_fe = all_hashed_images
-                                .get(parent_hash)
-                                .unwrap()
-                                .iter()
-                                .filter(|e| is_in_reference_folder(&self.directories.reference_directories, &e.path))
-                                .cloned()
-                                .collect();
-                            collected_similar_images.insert(parent_hash.clone(), vec_fe);
-                        }
-                    }
-
-                    for (child_hash, (parent_hash, similarity)) in hashes_similarity {
-                        let mut vec_fe: Vec<_> = all_hashed_images
-                            .get(child_hash)
-                            .unwrap()
-                            .iter()
-                            .filter(|e| !is_in_reference_folder(&self.directories.reference_directories, &e.path))
-                            .cloned()
-                            .collect();
-                        for mut fe in &mut vec_fe {
-                            fe.similarity = similarity;
-                        }
-                        collected_similar_images.get_mut(parent_hash).unwrap().append(&mut vec_fe);
-                    }
-                } else {
-                    // Collecting results to vector
-                    for (parent_hash, child_number) in hashes_parents {
-                        // If hash contains other hasher OR multiple images are available for checked hash
-                        if child_number > 0 || hashes_with_multiple_images.contains(parent_hash) {
-                            let vec_fe = all_hashed_images.get(parent_hash).unwrap().clone();
-                            collected_similar_images.insert(parent_hash.clone(), vec_fe);
-                        }
-                    }
-
-                    for (child_hash, (parent_hash, similarity)) in hashes_similarity {
-                        let mut vec_fe = all_hashed_images.get(child_hash).unwrap().clone();
-                        for mut fe in &mut vec_fe {
-                            fe.similarity = similarity;
-                        }
-                        collected_similar_images.get_mut(parent_hash).unwrap().append(&mut vec_fe);
-                    }
-                }
-            }
         }
 
-        // Validating if group contains duplicated results
-        #[cfg(debug_assertions)]
-        {
-            let mut result_hashset: HashSet<String> = Default::default();
-            let mut found = false;
-            for vec_file_entry in collected_similar_images.values() {
-                if vec_file_entry.is_empty() {
-                    println!("Empty Element {vec_file_entry:?}");
-                    found = true;
-                    continue;
-                }
-                if vec_file_entry.len() == 1 {
-                    println!("Single Element {vec_file_entry:?}");
-                    found = true;
-                    continue;
-                }
-                for file_entry in vec_file_entry {
-                    let st = file_entry.path.to_string_lossy().to_string();
-                    if result_hashset.contains(&st) {
-                        found = true;
-                        println!("Duplicated Element {st}");
-                    } else {
-                        result_hashset.insert(st);
-                    }
-                }
-            }
-            assert!(!found, "Found Invalid entries");
-        }
+        self.verify_duplicated_items(&collected_similar_images);
+
+        // Info about hashes is not needed anymore, so we drop this info
         self.similar_vectors = collected_similar_images.into_values().collect();
 
-        if self.exclude_images_with_same_size {
-            let mut new_vector = Default::default();
-            mem::swap(&mut self.similar_vectors, &mut new_vector);
-            for vec_file_entry in new_vector {
-                let mut bt_sizes: BTreeSet<u64> = Default::default();
-                let mut vec_values = Vec::new();
-                for file_entry in vec_file_entry {
-                    if !bt_sizes.contains(&file_entry.size) {
-                        bt_sizes.insert(file_entry.size);
-                        vec_values.push(file_entry);
-                    }
-                }
-                if vec_values.len() > 1 {
-                    self.similar_vectors.push(vec_values);
-                }
-            }
-        }
+        self.exclude_items_with_same_size();
 
-        if self.use_reference_folders {
-            let mut similar_vector = Default::default();
-            mem::swap(&mut self.similar_vectors, &mut similar_vector);
-            let reference_directories = self.directories.reference_directories.clone();
-            self.similar_referenced_vectors = similar_vector
-                .into_iter()
-                .filter_map(|vec_file_entry| {
-                    let mut files_from_referenced_folders = Vec::new();
-                    let mut normal_files = Vec::new();
-                    for file_entry in vec_file_entry {
-                        if reference_directories.iter().any(|e| file_entry.path.starts_with(e)) {
-                            files_from_referenced_folders.push(file_entry);
-                        } else {
-                            normal_files.push(file_entry);
-                        }
-                    }
-
-                    if files_from_referenced_folders.is_empty() || normal_files.is_empty() {
-                        None
-                    } else {
-                        Some((files_from_referenced_folders.pop().unwrap(), normal_files))
-                    }
-                })
-                .collect::<Vec<(FileEntry, Vec<FileEntry>)>>();
-        }
-
-        Common::print_time(hash_map_modification, SystemTime::now(), "sort_images - selecting data from HashMap");
+        self.remove_multiple_records_from_reference_folders();
 
         if self.use_reference_folders {
             for (_fe, vector) in &self.similar_referenced_vectors {
@@ -1046,12 +846,83 @@ impl SimilarImages {
             }
         }
 
-        // Clean unused data
+        // Clean unused data to save ram
         self.image_hashes = Default::default();
         self.images_to_check = Default::default();
         self.bktree = BKTree::new(Hamming);
 
         true
+    }
+
+    fn exclude_items_with_same_size(&mut self) {
+        if self.exclude_images_with_same_size {
+            for vec_file_entry in mem::take(&mut self.similar_vectors) {
+                let mut bt_sizes: BTreeSet<u64> = Default::default();
+                let mut vec_values = Vec::new();
+                for file_entry in vec_file_entry {
+                    if !bt_sizes.contains(&file_entry.size) {
+                        bt_sizes.insert(file_entry.size);
+                        vec_values.push(file_entry);
+                    }
+                }
+                if vec_values.len() > 1 {
+                    self.similar_vectors.push(vec_values);
+                }
+            }
+        }
+    }
+
+    fn remove_multiple_records_from_reference_folders(&mut self) {
+        if self.use_reference_folders {
+            self.similar_referenced_vectors = mem::take(&mut self.similar_vectors)
+                .into_iter()
+                .filter_map(|vec_file_entry| {
+                    let (mut files_from_referenced_folders, normal_files): (Vec<_>, Vec<_>) =
+                        vec_file_entry.into_iter().partition(|e| self.directories.is_in_referenced_directory(e.get_path()));
+
+                    if files_from_referenced_folders.is_empty() || normal_files.is_empty() {
+                        None
+                    } else {
+                        Some((files_from_referenced_folders.pop().unwrap(), normal_files))
+                    }
+                })
+                .collect::<Vec<(FileEntry, Vec<FileEntry>)>>();
+        }
+    }
+
+    #[allow(dead_code)]
+    #[allow(unreachable_code)]
+    #[allow(unused_variables)]
+    // TODO this probably not works good when reference folders are used
+    pub fn verify_duplicated_items(&self, collected_similar_images: &HashMap<ImHash, Vec<FileEntry>>) {
+        #[cfg(not(debug_assertions))]
+        return;
+        // Validating if group contains duplicated results
+        let mut result_hashset: HashSet<String> = Default::default();
+        let mut found = false;
+        // dbg!(collected_similar_images.len());
+        for vec_file_entry in collected_similar_images.values() {
+            if vec_file_entry.is_empty() {
+                println!("Empty group");
+                found = true;
+                continue;
+            }
+            if vec_file_entry.len() == 1 {
+                println!("Single Element {vec_file_entry:?}");
+                found = true;
+                continue;
+            }
+            for file_entry in vec_file_entry {
+                let st = file_entry.path.to_string_lossy().to_string();
+                if result_hashset.contains(&st) {
+                    found = true;
+                    println!("Duplicated Element {st}");
+                } else {
+                    result_hashset.insert(st);
+                }
+            }
+        }
+        assert!(!found, "Found Invalid entries, verify errors before"); // TODO crashes with empty result with reference folder, verify why
     }
 
     /// Set included dir which needs to be relative, exists etc.
@@ -1069,61 +940,6 @@ impl SimilarImages {
 
     pub fn set_excluded_items(&mut self, excluded_items: Vec<String>) {
         self.excluded_items.set_excluded_items(excluded_items, &mut self.text_messages);
-    }
-}
-
-fn image_to_check<'a>(
-    hashes_parents: &mut HashMap<&'a Vec<u8>, u32>,
-    hashes_similarity: &mut HashMap<&'a Vec<u8>, (&'a Vec<u8>, u32)>,
-    hashes_with_multiple_images: &HashSet<&'a Vec<u8>>,
-    hash_to_check: &'a Vec<u8>,
-    compared_hash: &'a Vec<u8>,
-    similarity: u32,
-) {
-    if let Some(children_number) = hashes_parents.get(compared_hash) {
-        if *children_number > 0 || hashes_with_multiple_images.contains(compared_hash) {
-            return;
-        }
-    }
-
-    // If there is already record, with smaller sensitivity, then replace it
-    let mut need_to_add = false;
-    let mut need_to_check = false;
-
-    // TODO consider to replace variables from above with closures
-    // If current checked hash, have parent, first we must check if similarity between them is lower than checked item
-    if let Some((current_parent_hash, current_similarity_with_parent)) = hashes_similarity.get(hash_to_check) {
-        if *current_similarity_with_parent > similarity {
-            need_to_check = true;
-
-            *hashes_parents.get_mut(current_parent_hash).unwrap() -= 1;
-            hashes_similarity.remove(hash_to_check).unwrap();
-        }
-    } else {
-        need_to_check = true;
-    }
-
-    if need_to_check {
-        if let Some((other_parent_hash, other_similarity)) = hashes_similarity.get(compared_hash) {
-            if *other_similarity > similarity {
-                need_to_add = true;
-                *hashes_parents.get_mut(other_parent_hash).unwrap() -= 1;
-            }
-        }
-        // But when there is no record, just add it
-        else {
-            need_to_add = true;
-        }
-    }
-
-    if need_to_add {
-        hashes_similarity.insert(compared_hash, (hash_to_check, similarity));
-
-        if let Some(number_of_children) = hashes_parents.get_mut(hash_to_check) {
-            *number_of_children += 1;
-        } else {
-            hashes_parents.insert(hash_to_check, 1);
-        }
     }
 }
 
@@ -1154,7 +970,6 @@ impl DebugPrint for SimilarImages {
 
 impl SaveResults for SimilarImages {
     fn save_results_to_file(&mut self, file_name: &str) -> bool {
-        let start_time: SystemTime = SystemTime::now();
         let file_name: String = match file_name {
             "" => "results.txt".to_string(),
             k => k.to_string(),
@@ -1200,7 +1015,6 @@ impl SaveResults for SimilarImages {
             write!(writer, "Not found any similar images.").unwrap();
         }
 
-        Common::print_time(start_time, SystemTime::now(), "save_results_to_file");
         true
     }
 }
@@ -1325,28 +1139,8 @@ pub fn get_string_from_similarity(similarity: &u32, hash_size: u8) -> String {
         16 => 1,
         32 => 2,
         64 => 3,
-        _ => panic!(),
+        _ => panic!("Invalid hash size {hash_size}"),
     };
-
-    // #[cfg(debug_assertions)]
-    // {
-    //     if *similarity <= SIMILAR_VALUES[index_preset][0] {
-    //         format!("{} {}", flc!("core_similarity_very_high"), *similarity)
-    //     } else if *similarity <= SIMILAR_VALUES[index_preset][1] {
-    //         format!("{} {}", flc!("core_similarity_high"), *similarity)
-    //     } else if *similarity <= SIMILAR_VALUES[index_preset][2] {
-    //         format!("{} {}", flc!("core_similarity_medium"), *similarity)
-    //     } else if *similarity <= SIMILAR_VALUES[index_preset][3] {
-    //         format!("{} {}", flc!("core_similarity_small"), *similarity)
-    //     } else if *similarity <= SIMILAR_VALUES[index_preset][4] {
-    //         format!("{} {}", flc!("core_similarity_very_small"), *similarity)
-    //     } else if *similarity <= SIMILAR_VALUES[index_preset][5] {
-    //         format!("{} {}", flc!("core_similarity_minimal"), *similarity)
-    //     } else {
-    //         panic!();
-    //     }
-    // }
-    // #[cfg(not(debug_assertions))]
 
     if *similarity == 0 {
         flc!("core_similarity_original")
@@ -1452,27 +1246,37 @@ pub fn test_image_conversion_speed() {
 }
 
 #[allow(dead_code)]
+#[allow(unreachable_code)]
+#[allow(unused_variables)]
 // Function to validate if after first check there are any duplicated entries
 // E.g. /a.jpg is used also as master and similar image which is forbidden, because may
 // cause accidentally delete more pictures that user wanted
 fn debug_check_for_duplicated_things(
-    hashes_parents: &HashMap<&Vec<u8>, u32>,
-    hashes_similarity: &HashMap<&Vec<u8>, (&Vec<u8>, u32)>,
-    all_hashed_images: &HashMap<Vec<u8>, Vec<FileEntry>>,
+    use_reference_folders: bool,
+    hashes_parents: &HashMap<ImHash, u32>,
+    hashes_similarity: &HashMap<ImHash, (ImHash, u32)>,
+    all_hashed_images: &HashMap<ImHash, Vec<FileEntry>>,
     numm: &str,
 ) {
+    #[cfg(not(debug_assertions))]
+    return;
+
+    if use_reference_folders {
+        return;
+    }
+
     let mut found_broken_thing = false;
     let mut hashmap_hashes: HashSet<_> = Default::default();
     let mut hashmap_names: HashSet<_> = Default::default();
     for (hash, number_of_children) in hashes_parents {
         if *number_of_children > 0 {
-            if hashmap_hashes.contains(*hash) {
-                println!("------1--HASH--{}  {:?}", numm, all_hashed_images.get(*hash).unwrap());
+            if hashmap_hashes.contains(hash) {
+                println!("------1--HASH--{}  {:?}", numm, all_hashed_images.get(hash).unwrap());
                 found_broken_thing = true;
             }
             hashmap_hashes.insert((*hash).clone());
 
-            for i in all_hashed_images.get(*hash).unwrap() {
+            for i in all_hashed_images.get(hash).unwrap() {
                 let name = i.path.to_string_lossy().to_string();
                 if hashmap_names.contains(&name) {
                     println!("------1--NAME--{numm}  {name:?}");
@@ -1483,13 +1287,13 @@ fn debug_check_for_duplicated_things(
         }
     }
     for hash in hashes_similarity.keys() {
-        if hashmap_hashes.contains(*hash) {
-            println!("------2--HASH--{}  {:?}", numm, all_hashed_images.get(*hash).unwrap());
+        if hashmap_hashes.contains(hash) {
+            println!("------2--HASH--{}  {:?}", numm, all_hashed_images.get(hash).unwrap());
             found_broken_thing = true;
         }
         hashmap_hashes.insert((*hash).clone());
 
-        for i in all_hashed_images.get(*hash).unwrap() {
+        for i in all_hashed_images.get(hash).unwrap() {
             let name = i.path.to_string_lossy().to_string();
             if hashmap_names.contains(&name) {
                 println!("------2--NAME--{numm}  {name:?}");
@@ -1499,7 +1303,403 @@ fn debug_check_for_duplicated_things(
         }
     }
 
-    if found_broken_thing {
-        panic!();
+    assert!(!found_broken_thing);
+}
+
+#[cfg(test)]
+mod tests {
+    use bk_tree::BKTree;
+    use std::collections::HashMap;
+    use std::path::PathBuf;
+
+    use crate::common_directory::Directories;
+    use crate::similar_images::{FileEntry, Hamming, ImHash, SimilarImages};
+
+    #[test]
+    fn test_compare_no_images() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages::default();
+            similar_images.find_similar_images(None, None);
+            assert_eq!(similar_images.get_similar_images().len(), 0);
+        }
+    }
+
+    #[test]
+    fn test_compare_tolerance_0_normal_mode() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 0,
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "bcd.txt");
+            let fe3 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 2], "cde.txt");
+            let fe4 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 2], "rrt.txt");
+            let fe5 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 2], "bld.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1.clone(), fe2.clone(), fe3.clone(), fe4.clone(), fe5.clone()]);
+
+            similar_images.find_similar_hashes(None, None);
+            assert_eq!(similar_images.get_similar_images().len(), 2);
+            let first_group = similar_images.get_similar_images()[0].iter().map(|e| &e.path).collect::<Vec<_>>();
+            let second_group = similar_images.get_similar_images()[1].iter().map(|e| &e.path).collect::<Vec<_>>();
+            // Initial order is not guaranteed, so we need to check both options
+            if similar_images.get_similar_images()[0][0].hash == fe1.hash {
+                assert_eq!(first_group, vec![&fe1.path, &fe2.path]);
+                assert_eq!(second_group, vec![&fe3.path, &fe4.path, &fe5.path]);
+            } else {
+                assert_eq!(first_group, vec![&fe3.path, &fe4.path, &fe5.path]);
+                assert_eq!(second_group, vec![&fe1.path, &fe2.path]);
+            }
+        }
+    }
+
+    #[test]
+    fn test_simple_normal_one_group() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 1,
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "bcd.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2]);
+
+            similar_images.find_similar_hashes(None, None);
+            assert_eq!(similar_images.get_similar_images().len(), 1);
+        }
+    }
+
+    #[test]
+    fn test_simple_normal_one_group_extended() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 2,
+                use_reference_folders: false,
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "bcd.txt");
+            let fe3 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 2], "rrd.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2, fe3]);
+
+            similar_images.find_similar_hashes(None, None);
+            assert_eq!(similar_images.get_similar_images().len(), 1);
+            assert_eq!(similar_images.get_similar_images()[0].len(), 3);
+        }
+    }
+
+    // TODO this not works yet,
+    // Need to find a way to
+    // #[test]
+    // fn test_similar_similarity() {
+    //     for _ in 0..100 {
+    //         let mut similar_images = SimilarImages {
+    //             similarity: 10,
+    //             use_reference_folders: false,
+    //             ..Default::default()
+    //         };
+    //
+    //         let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0000_0001], "abc.txt");
+    //         let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0000_0010], "bcd.txt");
+    //         let fe3 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0000_0100], "rrd.txt");
+    //         let fe4 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0111_1111], "rdd.txt");
+    //
+    //         add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2, fe3, fe4]);
+    //
+    //         similar_images.find_similar_hashes(None, None);
+    //         assert_eq!(similar_images.get_similar_images().len(), 1);
+    //         assert_eq!(similar_images.get_similar_images()[0].len(), 4);
+    //     }
+    // }
+
+    #[test]
+    fn test_simple_referenced_same_group() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 0,
+                use_reference_folders: true,
+                directories: Directories {
+                    reference_directories: vec![PathBuf::from("/home/rr/")],
+                    ..Default::default()
+                },
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "/home/rr/abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "/home/rr/bcd.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2]);
+
+            similar_images.find_similar_hashes(None, None);
+            assert_eq!(similar_images.get_similar_images().len(), 0);
+        }
+    }
+
+    #[test]
+    fn test_simple_referenced_group_extended() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 0,
+                use_reference_folders: true,
+                directories: Directories {
+                    reference_directories: vec![PathBuf::from("/home/rr/")],
+                    ..Default::default()
+                },
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "/home/rr/abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "/home/kk/bcd.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2]);
+
+            similar_images.find_similar_hashes(None, None);
+            assert_eq!(similar_images.get_similar_images_referenced().len(), 1);
+            assert_eq!(similar_images.get_similar_images_referenced()[0].1.len(), 1);
+        }
+    }
+
+    #[test]
+    fn test_simple_referenced_group_extended2() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 0,
+                use_reference_folders: true,
+                directories: Directories {
+                    reference_directories: vec![PathBuf::from("/home/rr/")],
+                    ..Default::default()
+                },
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "/home/rr/abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "/home/rr/abc2.txt");
+            let fe3 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "/home/kk/bcd.txt");
+            let fe4 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 1], "/home/kk/bcd2.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2, fe3, fe4]);
+
+            similar_images.find_similar_hashes(None, None);
+            let res = similar_images.get_similar_images_referenced();
+            assert_eq!(res.len(), 1);
+            assert_eq!(res[0].1.len(), 2);
+            assert!(res[0].1.iter().all(|e| e.path.starts_with("/home/kk/")));
+        }
+    }
+
+    #[test]
+    fn test_simple_normal_too_small_similarity() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 1,
+                use_reference_folders: false,
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b00001], "abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b00100], "bcd.txt");
+            let fe3 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b10000], "rrd.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2, fe3]);
+
+            similar_images.find_similar_hashes(None, None);
+            let res = similar_images.get_similar_images();
+            // dbg!(&res);
+            assert!(res.is_empty());
+        }
+    }
+
+    #[test]
+    fn test_simple_normal_union_of_similarity() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 4,
+                use_reference_folders: false,
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0000_0001], "abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0000_1111], "bcd.txt");
+            let fe3 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0111_1111], "rrd.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2, fe3]);
+
+            similar_images.find_similar_hashes(None, None);
+            let res = similar_images.get_similar_images();
+            assert_eq!(res.len(), 1);
+            let mut path = res[0].iter().map(|e| e.path.to_string_lossy().to_string()).collect::<Vec<_>>();
+            path.sort();
+            if res[0].len() == 3 {
+                assert_eq!(path, vec!["abc.txt".to_string(), "bcd.txt".to_string(), "rrd.txt".to_string()]);
+            } else if res[0].len() == 2 {
+                assert!(path == vec!["abc.txt".to_string(), "bcd.txt".to_string()] || path == vec!["bcd.txt".to_string(), "rrd.txt".to_string()]);
+            } else {
+                panic!("Invalid number of items");
+            }
+        }
+    }
+    #[test]
+    fn test_reference_similarity_only_one() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 1,
+                use_reference_folders: true,
+                directories: Directories {
+                    reference_directories: vec![PathBuf::from("/home/rr/")],
+                    ..Default::default()
+                },
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0001], "/home/rr/abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0011], "/home/kk/bcd.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2]);
+
+            similar_images.find_similar_hashes(None, None);
+            let res = similar_images.get_similar_images_referenced();
+            assert_eq!(res.len(), 1);
+            assert_eq!(res[0].1.len(), 1);
+            assert_eq!(res[0].0.path, PathBuf::from("/home/rr/abc.txt"));
+            assert_eq!(res[0].1[0].path, PathBuf::from("/home/kk/bcd.txt"));
+        }
+    }
+    #[test]
+    fn test_reference_too_small_similarity() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 1,
+                use_reference_folders: true,
+                directories: Directories {
+                    reference_directories: vec![PathBuf::from("/home/rr/")],
+                    ..Default::default()
+                },
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0001], "/home/rr/abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0010], "/home/kk/bcd.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2]);
+
+            similar_images.find_similar_hashes(None, None);
+            let res = similar_images.get_similar_images_referenced();
+            assert_eq!(res.len(), 0);
+        }
+    }
+
+    #[test]
+    fn test_reference_minimal() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 1,
+                use_reference_folders: true,
+                directories: Directories {
+                    reference_directories: vec![PathBuf::from("/home/rr/")],
+                    ..Default::default()
+                },
+                ..Default::default()
+            };
+
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0001], "/home/rr/abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0011], "/home/kk/bcd.txt");
+            let fe3 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0100], "/home/kk/bcd2.txt");
+            let fe4 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b1100], "/home/rr/krkr.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe1, fe2, fe3, fe4]);
+
+            similar_images.find_similar_hashes(None, None);
+            let res = similar_images.get_similar_images_referenced();
+            assert_eq!(res.len(), 2);
+            assert_eq!(res[0].1.len(), 1);
+            assert_eq!(res[1].1.len(), 1);
+            if res[0].1[0].path == PathBuf::from("/home/kk/bcd.txt") {
+                assert_eq!(res[0].0.path, PathBuf::from("/home/rr/abc.txt"));
+                assert_eq!(res[1].0.path, PathBuf::from("/home/rr/krkr.txt"));
+            } else if res[0].1[0].path == PathBuf::from("/home/kk/bcd2.txt") {
+                assert_eq!(res[0].0.path, PathBuf::from("/home/rr/krkr.txt"));
+                assert_eq!(res[1].0.path, PathBuf::from("/home/rr/abc.txt"));
+            }
+        }
+    }
+
+    #[test]
+    fn test_reference_union() {
+        for _ in 0..100 {
+            let mut similar_images = SimilarImages {
+                similarity: 10,
+                use_reference_folders: true,
+                directories: Directories {
+                    reference_directories: vec![PathBuf::from("/home/rr/")],
+                    ..Default::default()
+                },
+                ..Default::default()
+            };
+
+            let fe0 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b1000], "/home/rr/abc2.txt");
+            let fe1 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0001], "/home/rr/abc.txt");
+            let fe2 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b1110], "/home/kk/bcd.txt");
+            let fe3 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b0100], "/home/kk/bcd2.txt");
+            let fe4 = create_random_file_entry(vec![1, 1, 1, 1, 1, 1, 1, 0b1100], "/home/rr/krkr.txt");
+
+            add_hashes(&mut similar_images.image_hashes, vec![fe0, fe1, fe2, fe3, fe4]);
+
+            similar_images.find_similar_hashes(None, None);
+            let res = similar_images.get_similar_images_referenced();
+            assert_eq!(res.len(), 1);
+            assert_eq!(res[0].1.len(), 2);
+            assert_eq!(res[0].0.path, PathBuf::from("/home/rr/krkr.txt"));
+        }
+    }
+
+    #[test]
+    fn test_tolerance() {
+        // This test not really tests anything, but shows that current hamming distance works
+        // in bits instead of bytes
+        // I tried to make it work in bytes, but it was terrible, so Hamming should be really Ok
+
+        let fe1 = vec![1, 1, 1, 1, 1, 1, 1, 1];
+        let fe2 = vec![1, 1, 1, 1, 1, 1, 1, 2];
+        let mut bktree = BKTree::new(Hamming);
+        bktree.add(fe1);
+        let (similarity, _hash) = bktree.find(&fe2, 100).next().unwrap();
+        assert_eq!(similarity, 2);
+
+        let fe1 = vec![1, 1, 1, 1, 1, 1, 1, 1];
+        let fe2 = vec![1, 1, 1, 1, 1, 1, 1, 3];
+        let mut bktree = BKTree::new(Hamming);
+        bktree.add(fe1);
+        let (similarity, _hash) = bktree.find(&fe2, 100).next().unwrap();
+        assert_eq!(similarity, 1);
+
+        let fe1 = vec![1, 1, 1, 1, 1, 1, 1, 0b0000_0000];
+        let fe2 = vec![1, 1, 1, 1, 1, 1, 1, 0b0000_1000];
+        let mut bktree = BKTree::new(Hamming);
+        bktree.add(fe1);
+        let (similarity, _hash) = bktree.find(&fe2, 100).next().unwrap();
+        assert_eq!(similarity, 1);
+    }
+
+    fn add_hashes(hashmap: &mut HashMap<ImHash, Vec<FileEntry>>, file_entries: Vec<FileEntry>) {
+        for fe in file_entries {
+            hashmap.entry(fe.hash.clone()).or_insert_with(Vec::new).push(fe);
+        }
+    }
+
+    fn create_random_file_entry(hash: Vec<u8>, name: &str) -> FileEntry {
+        FileEntry {
+            path: PathBuf::from(name.to_string()),
+            size: 0,
+            dimensions: String::new(),
+            modified_date: 0,
+            hash,
+            similarity: 0,
+        }
     }
 }
