@@ -29,7 +29,7 @@ pub struct CheckMethodStruct {
     pub check_method: CheckingMethod,
 }
 
-pub const DUPLICATES_CHECK_METHOD_COMBO_BOX: [CheckMethodStruct; 3] = [
+pub const DUPLICATES_CHECK_METHOD_COMBO_BOX: [CheckMethodStruct; 4] = [
     CheckMethodStruct {
         eng_name: "Hash",
         check_method: CheckingMethod::Hash,
@@ -41,6 +41,27 @@ pub const DUPLICATES_CHECK_METHOD_COMBO_BOX: [CheckMethodStruct; 3] = [
     CheckMethodStruct {
         eng_name: "Name",
         check_method: CheckingMethod::Name,
+    },
+    CheckMethodStruct {
+        eng_name: "Size and Name",
+        check_method: CheckingMethod::SizeName,
+    },
+];
+
+#[derive(Copy, Clone)]
+pub struct AudioTypeStruct {
+    pub eng_name: &'static str,
+    pub check_method: CheckingMethod,
+}
+
+pub const AUDIO_TYPE_CHECK_METHOD_COMBO_BOX: [AudioTypeStruct; 2] = [
+    AudioTypeStruct {
+        eng_name: "Tags",
+        check_method: CheckingMethod::AudioTags,
+    },
+    AudioTypeStruct {
+        eng_name: "Content",
+        check_method: CheckingMethod::AudioContent,
     },
 ];
 
